@@ -1,206 +1,310 @@
 {
-  "name": "element-web",
-  "version": "1.10.13",
-  "description": "A feature-rich client for Matrix.org",
-  "author": "New Vector Ltd.",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/vector-im/element-web"
-  },
-  "license": "Apache-2.0",
-  "files": [
-    "lib",
-    "res",
-    "src",
-    "webpack.config.js",
-    "scripts",
-    "docs",
-    "release.sh",
-    "deploy",
-    "CHANGELOG.md",
-    "CONTRIBUTING.rst",
-    "LICENSE",
-    "README.md",
-    "AUTHORS.rst",
-    "package.json",
-    "contribute.json"
-  ],
-  "style": "bundle.css",
-  "scripts": {
-    "i18n": "matrix-gen-i18n",
-    "prunei18n": "matrix-prune-i18n",
-    "diff-i18n": "cp src/i18n/strings/en_EN.json src/i18n/strings/en_EN_orig.json && matrix-gen-i18n && matrix-compare-i18n-files src/i18n/strings/en_EN_orig.json src/i18n/strings/en_EN.json",
-    "clean": "rimraf lib webapp",
-    "build": "yarn clean && yarn build:genfiles && yarn build:bundle",
-    "build-stats": "yarn clean && yarn build:genfiles && yarn build:bundle-stats",
-    "build:jitsi": "node scripts/build-jitsi.js",
-    "build:res": "node scripts/copy-res.js",
-    "build:genfiles": "yarn build:res && yarn build:jitsi",
-    "build:modernizr": "modernizr -c .modernizr.json -d src/vector/modernizr.js",
-    "build:bundle": "webpack --progress --bail --mode production",
-    "build:bundle-stats": "webpack --progress --bail --mode production --json > webpack-stats.json",
-    "dist": "scripts/package.sh",
-    "start": "concurrently --kill-others-on-fail --prefix \"{time} [{name}]\" -n res,element-js \"yarn start:res\" \"yarn start:js\"",
-    "start:https": "concurrently --kill-others-on-fail --prefix \"{time} [{name}]\" -n res,element-js \"yarn start:res\" \"yarn start:js --https\"",
-    "start:res": "yarn build:jitsi && node scripts/copy-res.js -w",
-    "start:js": "webpack-dev-server --host=0.0.0.0 --output-filename=bundles/_dev_/[name].js --output-chunk-filename=bundles/_dev_/[name].js -w --mode development --disable-host-check --hot",
-    "lint": "yarn lint:types && yarn lint:js && yarn lint:style",
-    "lint:js": "eslint --max-warnings 0 src",
-    "lint:js-fix": "eslint --fix src",
-    "lint:types": "tsc --noEmit --jsx react",
-    "lint:style": "stylelint \"res/css/**/*.scss\"",
-    "test": "jest",
-    "coverage": "yarn test --coverage"
-  },
-  "dependencies": {
-    "@matrix-org/olm": "https://gitlab.matrix.org/api/v4/projects/27/packages/npm/@matrix-org/olm/-/@matrix-org/olm-3.2.8.tgz",
-    "browser-request": "^0.3.3",
-    "gfm.css": "^1.1.2",
-    "jsrsasign": "^10.2.0",
-    "katex": "^0.12.0",
-    "matrix-js-sdk": "github:matrix-org/matrix-js-sdk#develop",
-    "matrix-react-sdk": "github:matrix-org/matrix-react-sdk#develop",
-    "matrix-widget-api": "^0.1.0-beta.18",
-    "prop-types": "^15.7.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "sanitize-html": "^2.3.2",
-    "ua-parser-js": "^0.7.24"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/eslint-parser": "^7.12.10",
-    "@babel/eslint-plugin": "^7.12.10",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/plugin-proposal-export-default-from": "^7.12.1",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
-    "@babel/plugin-proposal-numeric-separator": "^7.12.7",
-    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
-    "@babel/plugin-proposal-optional-chaining": "^7.12.7",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-react": "^7.12.10",
-    "@babel/preset-typescript": "^7.12.7",
-    "@babel/register": "^7.12.10",
-    "@babel/runtime": "^7.12.5",
-    "@principalstudio/html-webpack-inject-preload": "^1.2.7",
-    "@sentry/webpack-plugin": "^1.18.1",
-    "@svgr/webpack": "^5.5.0",
-    "@types/flux": "^3.1.9",
-    "@types/jest": "^27.0.2",
-    "@types/modernizr": "^3.5.3",
-    "@types/node": "^14.14.22",
-    "@types/react": "17.0.14",
-    "@types/react-dom": "17.0.9",
-    "@types/sanitize-html": "^2.3.1",
-    "@types/ua-parser-js": "^0.7.36",
-    "@typescript-eslint/eslint-plugin": "^5.6.0",
-    "@typescript-eslint/parser": "^5.6.0",
-    "allchange": "^1.0.6",
-    "autoprefixer": "^9.8.6",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.2",
-    "chokidar": "^3.5.1",
-    "concurrently": "^5.3.0",
-    "cpx": "^1.5.0",
-    "css-loader": "^3.6.0",
-    "dotenv": "^10.0.0",
-    "eslint": "8.9.0",
-    "eslint-config-google": "^0.14.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-matrix-org": "^0.4.0",
-    "eslint-plugin-react": "^7.28.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "fake-indexeddb": "^3.1.2",
-    "file-loader": "^5.1.0",
-    "fs-extra": "^0.30.0",
-    "html-webpack-plugin": "^4.5.2",
-    "jest": "^26.6.3",
-    "jest-environment-jsdom-sixteen": "^1.0.3",
-    "jest-raw-loader": "^1.0.1",
-    "jest-sonar-reporter": "^2.0.0",
-    "json-loader": "^0.5.7",
-    "loader-utils": "^1.4.0",
-    "matrix-mock-request": "^2.0.0",
-    "matrix-react-test-utils": "^0.2.3",
-    "matrix-web-i18n": "^1.2.0",
-    "mini-css-extract-plugin": "^0.12.0",
-    "minimist": "^1.2.6",
-    "mkdirp": "^1.0.4",
-    "modernizr": "^3.12.0",
-    "node-fetch": "^2.6.7",
-    "optimize-css-assets-webpack-plugin": "^5.0.4",
-    "postcss-easings": "^2.0.0",
-    "postcss-hexrgba": "^2.0.1",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-mixins": "^6.2.3",
-    "postcss-nested": "^4.2.3",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-scss": "^2.1.1",
-    "postcss-simple-vars": "^5.0.2",
-    "postcss-strip-inline-comments": "^0.1.5",
-    "raw-loader": "^4.0.2",
-    "rimraf": "^3.0.2",
-    "shell-escape": "^0.2.0",
-    "simple-proxy-agent": "^1.1.0",
-    "string-replace-loader": "2",
-    "style-loader": "2",
-    "stylelint": "^13.9.0",
-    "stylelint-config-standard": "^20.0.0",
-    "stylelint-scss": "^3.18.0",
-    "terser-webpack-plugin": "^2.3.8",
-    "typescript": "^4.5.3",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.2",
-    "worker-loader": "^2.0.0",
-    "worklet-loader": "^2.0.0"
-  },
-  "resolutions": {
-    "@types/react": "17.0.14"
-  },
-  "jest": {
-    "testEnvironment": "jest-environment-jsdom-sixteen",
-    "testMatch": [
-      "<rootDir>/test/**/*-test.[tj]s?(x)"
-    ],
-    "setupFilesAfterEnv": [
-      "<rootDir>/node_modules/matrix-react-sdk/test/setupTests.js"
-    ],
-    "moduleNameMapper": {
-      "\\.(css|scss)$": "<rootDir>/__mocks__/cssMock.js",
-      "\\.(gif|png|ttf|woff2)$": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/imageMock.js",
-      "\\.svg$": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/svg.js",
-      "\\$webapp/i18n/languages.json": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/languages.json",
-      "^browser-request$": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/browser-request.js",
-      "^react$": "<rootDir>/node_modules/react",
-      "^react-dom$": "<rootDir>/node_modules/react-dom",
-      "^matrix-js-sdk$": "<rootDir>/node_modules/matrix-js-sdk/src",
-      "^matrix-react-sdk$": "<rootDir>/node_modules/matrix-react-sdk/src",
-      "decoderWorker\\.min\\.js": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/empty.js",
-      "decoderWorker\\.min\\.wasm": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/empty.js",
-      "waveWorker\\.min\\.js": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/empty.js",
-      "context-filter-polyfill": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/empty.js",
-      "FontManager.ts": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/FontManager.js",
-      "workers/(.+)\\.worker\\.ts": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/workerMock.js",
-      "^!!raw-loader!.*": "jest-raw-loader",
-      "RecorderWorklet": "<rootDir>/node_modules/matrix-react-sdk/__mocks__/empty.js"
+    "name": "element-web",
+    "version": "1.11.91",
+    "description": "Element: the future of secure communication",
+    "author": "New Vector Ltd.",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/element-hq/element-web"
     },
-    "transformIgnorePatterns": [
-      "/node_modules/(?!matrix-js-sdk).+$",
-      "/node_modules/(?!matrix-react-sdk).+$"
+    "license": "SEE LICENSE IN README.md",
+    "files": [
+        "lib",
+        "res",
+        "src",
+        "webpack.config.js",
+        "scripts",
+        "docs",
+        "release.sh",
+        "deploy",
+        "CHANGELOG.md",
+        "CONTRIBUTING.rst",
+        "LICENSE",
+        "README.md",
+        "AUTHORS.rst",
+        "package.json",
+        "contribute.json"
     ],
-    "coverageReporters": [
-      "text-summary",
-      "lcov"
+    "style": "bundle.css",
+    "matrix_i18n_extra_translation_funcs": [
+        "UserFriendlyError"
     ],
-    "testResultsProcessor": "jest-sonar-reporter"
-  },
-  "jestSonar": {
-    "reportPath": "coverage",
-    "sonar56x": true
-  }
+    "scripts": {
+        "i18n": "matrix-gen-i18n && yarn i18n:sort && yarn i18n:lint",
+        "i18n:sort": "jq --sort-keys '.' src/i18n/strings/en_EN.json > src/i18n/strings/en_EN.json.tmp && mv src/i18n/strings/en_EN.json.tmp src/i18n/strings/en_EN.json",
+        "i18n:lint": "matrix-i18n-lint && prettier --log-level=silent --write src/i18n/strings/ --ignore-path /dev/null",
+        "i18n:diff": "cp src/i18n/strings/en_EN.json src/i18n/strings/en_EN_orig.json && yarn i18n && matrix-compare-i18n-files src/i18n/strings/en_EN_orig.json src/i18n/strings/en_EN.json",
+        "make-component": "node scripts/make-react-component.js",
+        "rethemendex": "./res/css/rethemendex.sh",
+        "clean": "rimraf lib webapp",
+        "build": "yarn clean && yarn build:genfiles && yarn build:bundle",
+        "build-stats": "yarn clean && yarn build:genfiles && yarn build:bundle-stats",
+        "build:res": "ts-node scripts/copy-res.ts",
+        "build:genfiles": "yarn build:res && yarn build:module_system",
+        "build:modernizr": "modernizr -c .modernizr.json -d src/vector/modernizr.js",
+        "build:bundle": "webpack --progress --mode production",
+        "build:bundle-stats": "webpack --progress --mode production --json > webpack-stats.json",
+        "build:module_system": "ts-node --project ./tsconfig.module_system.json module_system/scripts/install.ts",
+        "dist": "./scripts/package.sh",
+        "start": "concurrently --kill-others-on-fail --prefix \"{time} [{name}]\" -n modules,res \"yarn build:module_system\" \"yarn build:res\" && concurrently --kill-others-on-fail --prefix \"{time} [{name}]\" -n res,element-js \"yarn start:res\" \"yarn start:js\"",
+        "start:https": "concurrently --kill-others-on-fail --prefix \"{time} [{name}]\" -n res,element-js \"yarn start:res\" \"yarn start:js --server-type https\"",
+        "start:res": "ts-node scripts/copy-res.ts -w",
+        "start:js": "webpack serve --output-path webapp --output-filename=bundles/_dev_/[name].js --output-chunk-filename=bundles/_dev_/[name].js --mode development",
+        "lint": "yarn lint:types && yarn lint:js && yarn lint:style && yarn lint:workflows",
+        "lint:js": "eslint --max-warnings 0 src test playwright module_system && prettier --check .",
+        "lint:js-fix": "prettier --log-level=warn --write . && eslint --fix src test playwright module_system",
+        "lint:types": "yarn lint:types:src && yarn lint:types:module_system",
+        "lint:types:src": "tsc --noEmit --jsx react && tsc --noEmit --jsx react -p playwright",
+        "lint:types:module_system": "tsc --noEmit --project ./tsconfig.module_system.json",
+        "lint:style": "stylelint \"res/css/**/*.pcss\"",
+        "lint:workflows": "find .github/workflows -type f \\( -iname '*.yaml' -o -iname '*.yml' \\) | xargs -I {} sh -c 'echo \"Linting {}\"; action-validator \"{}\"'",
+        "lint:knip": "knip",
+        "test": "jest",
+        "test:playwright": "playwright test",
+        "test:playwright:open": "yarn test:playwright --ui",
+        "test:playwright:screenshots": "yarn test:playwright:screenshots:build && yarn test:playwright:screenshots:run",
+        "test:playwright:screenshots:build": "docker build playwright -t element-web-playwright",
+        "test:playwright:screenshots:run": "docker run --rm --network host -e BASE_URL -e CI -v $(pwd):/work/ -v $(node -e 'console.log(require(`path`).dirname(require.resolve(`matrix-js-sdk/package.json`)))'):/work/node_modules/matrix-js-sdk -v /var/run/docker.sock:/var/run/docker.sock -v /tmp/:/tmp/ -it element-web-playwright --grep @screenshot --project=Chrome",
+        "coverage": "yarn test --coverage",
+        "analyse:webpack-bundles": "webpack-bundle-analyzer webpack-stats.json webapp",
+        "update:jitsi": "curl -s https://meet.element.io/libs/external_api.min.js > ./res/jitsi_external_api.min.js"
+    },
+    "resolutions": {
+        "@types/react": "18.3.18",
+        "@types/react-dom": "18.3.5",
+        "oidc-client-ts": "3.1.0",
+        "jwt-decode": "4.0.0",
+        "caniuse-lite": "1.0.30001697",
+        "wrap-ansi-cjs": "npm:wrap-ansi@^7.0.0",
+        "wrap-ansi": "npm:wrap-ansi@^7.0.0"
+    },
+    "dependencies": {
+        "@babel/runtime": "^7.12.5",
+        "@element-hq/element-web-module-api": "^0.1.1",
+        "@fontsource/inconsolata": "^5",
+        "@fontsource/inter": "^5",
+        "@formatjs/intl-segmenter": "^11.5.7",
+        "@matrix-org/analytics-events": "^0.29.0",
+        "@matrix-org/emojibase-bindings": "^1.3.4",
+        "@matrix-org/react-sdk-module-api": "^2.4.0",
+        "@matrix-org/spec": "^1.7.0",
+        "@sentry/browser": "^8.0.0",
+        "@types/png-chunks-extract": "^1.0.2",
+        "@types/react-virtualized": "^9.21.30",
+        "@vector-im/compound-design-tokens": "^3.0.0",
+        "@vector-im/compound-web": "^7.6.1",
+        "@vector-im/matrix-wysiwyg": "2.38.0",
+        "@zxcvbn-ts/core": "^3.0.4",
+        "@zxcvbn-ts/language-common": "^3.0.4",
+        "@zxcvbn-ts/language-en": "^3.0.2",
+        "await-lock": "^2.1.0",
+        "bloom-filters": "^3.0.3",
+        "blurhash": "^2.0.3",
+        "browserslist": "^4.23.2",
+        "classnames": "^2.2.6",
+        "commonmark": "^0.31.0",
+        "counterpart": "^0.18.6",
+        "css-tree": "^3.0.0",
+        "diff-dom": "^5.0.0",
+        "diff-match-patch": "^1.0.5",
+        "emojibase-regex": "15.3.2",
+        "escape-html": "^1.0.3",
+        "file-saver": "^2.0.5",
+        "filesize": "10.1.6",
+        "github-markdown-css": "^5.5.1",
+        "glob-to-regexp": "^0.4.1",
+        "highlight.js": "^11.3.1",
+        "html-entities": "^2.0.0",
+        "is-ip": "^3.1.0",
+        "js-xxhash": "^4.0.0",
+        "jsrsasign": "^11.0.0",
+        "jszip": "^3.7.0",
+        "katex": "^0.16.0",
+        "linkify-element": "4.2.0",
+        "linkify-react": "4.2.0",
+        "linkify-string": "4.2.0",
+        "linkifyjs": "4.2.0",
+        "lodash": "^4.17.21",
+        "maplibre-gl": "^5.0.0",
+        "matrix-encrypt-attachment": "^1.0.3",
+        "matrix-events-sdk": "0.0.1",
+        "matrix-js-sdk": "github:matrix-org/matrix-js-sdk#develop",
+        "matrix-widget-api": "^1.10.0",
+        "memoize-one": "^6.0.0",
+        "mime": "^4.0.4",
+        "oidc-client-ts": "^3.0.1",
+        "opus-recorder": "^8.0.3",
+        "pako": "^2.0.3",
+        "png-chunks-extract": "^1.0.0",
+        "posthog-js": "1.157.2",
+        "qrcode": "1.5.4",
+        "re-resizable": "6.10.3",
+        "react": "^18.3.1",
+        "react-beautiful-dnd": "^13.1.0",
+        "react-blurhash": "^0.3.0",
+        "react-dom": "^18.3.1",
+        "react-focus-lock": "^2.5.1",
+        "react-transition-group": "^4.4.1",
+        "react-virtualized": "^9.22.5",
+        "rfc4648": "^1.4.0",
+        "sanitize-filename": "^1.6.3",
+        "sanitize-html": "2.14.0",
+        "tar-js": "^0.3.0",
+        "temporal-polyfill": "^0.2.5",
+        "ua-parser-js": "^1.0.2",
+        "uuid": "^11.0.0",
+        "what-input": "^5.2.10"
+    },
+    "devDependencies": {
+        "@action-validator/cli": "^0.6.0",
+        "@action-validator/core": "^0.6.0",
+        "@axe-core/playwright": "^4.8.1",
+        "@babel/core": "^7.12.10",
+        "@babel/eslint-parser": "^7.12.10",
+        "@babel/eslint-plugin": "^7.12.10",
+        "@babel/plugin-proposal-export-default-from": "^7.12.1",
+        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+        "@babel/plugin-transform-class-properties": "^7.12.1",
+        "@babel/plugin-transform-logical-assignment-operators": "^7.20.7",
+        "@babel/plugin-transform-nullish-coalescing-operator": "^7.12.1",
+        "@babel/plugin-transform-numeric-separator": "^7.12.7",
+        "@babel/plugin-transform-object-rest-spread": "^7.12.1",
+        "@babel/plugin-transform-optional-chaining": "^7.12.7",
+        "@babel/plugin-transform-runtime": "^7.12.10",
+        "@babel/preset-env": "^7.12.11",
+        "@babel/preset-react": "^7.12.10",
+        "@babel/preset-typescript": "^7.12.7",
+        "@babel/runtime": "^7.12.5",
+        "@casualbot/jest-sonar-reporter": "2.2.7",
+        "@peculiar/webcrypto": "^1.4.3",
+        "@playwright/test": "^1.40.1",
+        "@principalstudio/html-webpack-inject-preload": "^1.2.7",
+        "@sentry/webpack-plugin": "^3.0.0",
+        "@stylistic/eslint-plugin": "^3.0.0",
+        "@svgr/webpack": "^8.0.0",
+        "@testcontainers/postgresql": "^10.16.0",
+        "@testing-library/dom": "^10.4.0",
+        "@testing-library/jest-dom": "^6.4.8",
+        "@testing-library/react": "^16.0.0",
+        "@testing-library/user-event": "^14.5.2",
+        "@types/commonmark": "^0.27.4",
+        "@types/counterpart": "^0.18.1",
+        "@types/css-tree": "^2.3.8",
+        "@types/diff-match-patch": "^1.0.32",
+        "@types/escape-html": "^1.0.1",
+        "@types/express": "^5.0.0",
+        "@types/file-saver": "^2.0.3",
+        "@types/glob-to-regexp": "^0.4.1",
+        "@types/jest": "29.5.12",
+        "@types/jitsi-meet": "^2.0.2",
+        "@types/jsrsasign": "^10.5.4",
+        "@types/katex": "^0.16.0",
+        "@types/lodash": "^4.14.168",
+        "@types/minimist": "^1.2.5",
+        "@types/modernizr": "^3.5.3",
+        "@types/node": "18",
+        "@types/node-fetch": "^2.6.2",
+        "@types/pako": "^2.0.0",
+        "@types/qrcode": "^1.3.5",
+        "@types/react": "18.3.18",
+        "@types/react-beautiful-dnd": "^13.0.0",
+        "@types/react-dom": "18.3.5",
+        "@types/react-transition-group": "^4.4.0",
+        "@types/sanitize-html": "2.13.0",
+        "@types/semver": "^7.5.8",
+        "@types/tar-js": "^0.3.5",
+        "@types/ua-parser-js": "^0.7.36",
+        "@types/uuid": "^10.0.0",
+        "@typescript-eslint/eslint-plugin": "^8.19.0",
+        "@typescript-eslint/parser": "^8.19.0",
+        "babel-jest": "^29.0.0",
+        "babel-loader": "^9.0.0",
+        "babel-plugin-jsx-remove-data-test-id": "^3.0.0",
+        "blob-polyfill": "^9.0.0",
+        "chokidar": "^4.0.0",
+        "concurrently": "^9.0.0",
+        "copy-webpack-plugin": "^12.0.0",
+        "core-js": "^3.38.1",
+        "cronstrue": "^2.41.0",
+        "css-loader": "^7.0.0",
+        "css-minimizer-webpack-plugin": "^7.0.0",
+        "dotenv": "^16.0.2",
+        "eslint": "8.57.1",
+        "eslint-config-google": "^0.14.0",
+        "eslint-config-prettier": "^10.0.0",
+        "eslint-plugin-deprecate": "0.8.5",
+        "eslint-plugin-import": "^2.25.4",
+        "eslint-plugin-jest": "^28.0.0",
+        "eslint-plugin-jsx-a11y": "^6.5.1",
+        "eslint-plugin-matrix-org": "^2.0.2",
+        "eslint-plugin-react": "^7.28.0",
+        "eslint-plugin-react-compiler": "^19.0.0-beta-df7b47d-20241124",
+        "eslint-plugin-react-hooks": "^5.0.0",
+        "eslint-plugin-unicorn": "^56.0.0",
+        "express": "^4.18.2",
+        "fake-indexeddb": "^6.0.0",
+        "fetch-mock": "9.11.0",
+        "fetch-mock-jest": "^1.5.1",
+        "file-loader": "^6.0.0",
+        "glob": "^11.0.0",
+        "html-webpack-plugin": "^5.5.3",
+        "husky": "^9.0.0",
+        "jest": "^29.6.2",
+        "jest-canvas-mock": "^2.5.2",
+        "jest-environment-jsdom": "^29.7.0",
+        "jest-mock": "^29.6.2",
+        "jest-raw-loader": "^1.0.1",
+        "jsqr": "^1.4.0",
+        "knip": "^5.36.2",
+        "lint-staged": "^15.0.2",
+        "mailpit-api": "^1.0.5",
+        "matrix-web-i18n": "^3.2.1",
+        "mini-css-extract-plugin": "2.9.2",
+        "minimist": "^1.2.6",
+        "modernizr": "^3.12.0",
+        "node-fetch": "^2.6.7",
+        "playwright-core": "^1.45.1",
+        "postcss": "8.4.46",
+        "postcss-easings": "^4.0.0",
+        "postcss-hexrgba": "2.1.0",
+        "postcss-import": "16.1.0",
+        "postcss-loader": "8.1.1",
+        "postcss-mixins": "^11.0.0",
+        "postcss-nested": "^7.0.0",
+        "postcss-preset-env": "^10.0.0",
+        "postcss-scss": "^4.0.4",
+        "postcss-simple-vars": "^7.0.1",
+        "prettier": "3.4.2",
+        "process": "^0.11.10",
+        "raw-loader": "^4.0.2",
+        "rimraf": "^6.0.0",
+        "semver": "^7.5.2",
+        "source-map-loader": "^5.0.0",
+        "strip-ansi": "^7.1.0",
+        "stylelint": "^16.13.0",
+        "stylelint-config-standard": "^37.0.0",
+        "stylelint-scss": "^6.0.0",
+        "stylelint-value-no-unknown-custom-properties": "^6.0.1",
+        "terser-webpack-plugin": "^5.3.9",
+        "testcontainers": "^10.16.0",
+        "ts-node": "^10.9.1",
+        "typescript": "5.7.3",
+        "util": "^0.12.5",
+        "web-streams-polyfill": "^4.0.0",
+        "webpack": "^5.89.0",
+        "webpack-bundle-analyzer": "^4.8.0",
+        "webpack-cli": "^6.0.0",
+        "webpack-dev-server": "^5.0.0",
+        "webpack-retry-chunk-load-plugin": "^3.1.1",
+        "webpack-version-file-plugin": "^0.5.0",
+        "yaml": "^2.3.3"
+    },
+    "@casualbot/jest-sonar-reporter": {
+        "outputDirectory": "coverage",
+        "outputName": "jest-sonar-report.xml",
+        "relativePaths": true
+    },
+    "engines": {
+        "node": ">=20.0.0"
+    }
 }
