@@ -1,47 +1,39 @@
 /*
-Copyright 2015, 2016 OpenMarket Ltd
-Copyright 2017 Vector Creations Ltd
+Copyright 2018-2024 New Vector Ltd.
 Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
-Copyright 2018 - 2021 New Vector Ltd
+Copyright 2017 Vector Creations Ltd
+Copyright 2015, 2016 OpenMarket Ltd
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
 */
 
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import olmWasmPath from "@matrix-org/olm/olm.wasm";
-import Olm from '@matrix-org/olm';
-import * as ReactDOM from "react-dom";
-import * as React from "react";
-import * as languageHandler from "matrix-react-sdk/src/languageHandler";
-import SettingsStore from "matrix-react-sdk/src/settings/SettingsStore";
-import PlatformPeg from "matrix-react-sdk/src/PlatformPeg";
-import SdkConfig from "matrix-react-sdk/src/SdkConfig";
-import { setTheme } from "matrix-react-sdk/src/theme";
+import { createRoot } from "react-dom/client";
+import React, { StrictMode } from "react";
 import { logger } from "matrix-js-sdk/src/logger";
+import { ModuleLoader } from "@element-hq/element-web-module-api";
 
+import type { QueryDict } from "matrix-js-sdk/src/utils";
+import * as languageHandler from "../languageHandler";
+import SettingsStore from "../settings/SettingsStore";
+import PlatformPeg from "../PlatformPeg";
+import SdkConfig from "../SdkConfig";
+import { setTheme } from "../theme";
+import { ModuleRunner } from "../modules/ModuleRunner";
+import type MatrixChat from "../components/structures/MatrixChat";
 import ElectronPlatform from "./platform/ElectronPlatform";
 import PWAPlatform from "./platform/PWAPlatform";
 import WebPlatform from "./platform/WebPlatform";
 import { initRageshake, initRageshakeStore } from "./rageshakesetup";
+import ModuleApi from "../modules/Api.ts";
 
 export const rageshakePromise = initRageshake();
 
-export function preparePlatform() {
+export function preparePlatform(): void {
     if (window.electron) {
         logger.log("Using Electron platform");
         PlatformPeg.set(new ElectronPlatform());
-    } else if (window.matchMedia('(display-mode: standalone)').matches) {
+    } else if (window.matchMedia("(display-mode: standalone)").matches) {
         logger.log("Using PWA platform");
         PlatformPeg.set(new PWAPlatform());
     } else {
@@ -50,7 +42,7 @@ export function preparePlatform() {
     }
 }
 
-export function setupLogStorage() {
+export function setupLogStorage(): Promise<void> {
     if (SdkConfig.get().bug_report_endpoint_url) {
         return initRageshakeStore();
     }
@@ -58,59 +50,22 @@ export function setupLogStorage() {
     return Promise.resolve();
 }
 
-export async function loadConfig() {
+export async function loadConfig(): Promise<void> {
     // XXX: We call this twice, once here and once in MatrixChat as a prop. We call it here to ensure
     // granular settings are loaded correctly and to avoid duplicating the override logic for the theme.
     //
     // Note: this isn't called twice for some wrappers, like the Jitsi wrapper.
-    const platformConfig = await PlatformPeg.get().getConfig();
+    const platformConfig = await PlatformPeg.get()?.getConfig();
     if (platformConfig) {
         SdkConfig.put(platformConfig);
     } else {
-        SdkConfig.unset(); // clears the config (sets to empty object)
+        SdkConfig.reset();
     }
 }
 
-export function loadOlm(): Promise<void> {
-    /* Load Olm. We try the WebAssembly version first, and then the legacy,
-     * asm.js version if that fails. For this reason we need to wait for this
-     * to finish before continuing to load the rest of the app. In future
-     * we could somehow pass a promise down to react-sdk and have it wait on
-     * that so olm can be loading in parallel with the rest of the app.
-     *
-     * We also need to tell the Olm js to look for its wasm file at the same
-     * level as index.html. It really should be in the same place as the js,
-     * ie. in the bundle directory, but as far as I can tell this is
-     * completely impossible with webpack. We do, however, use a hashed
-     * filename to avoid caching issues.
-     */
-    return Olm.init({
-        locateFile: () => olmWasmPath,
-    }).then(() => {
-        logger.log("Using WebAssembly Olm");
-    }).catch((e) => {
-        logger.log("Failed to load Olm: trying legacy version", e);
-        return new Promise((resolve, reject) => {
-            const s = document.createElement('script');
-            s.src = 'olm_legacy.js'; // XXX: This should be cache-busted too
-            s.onload = resolve;
-            s.onerror = reject;
-            document.body.appendChild(s);
-        }).then(() => {
-            // Init window.Olm, ie. the one just loaded by the script tag,
-            // not 'Olm' which is still the failed wasm version.
-            return window.Olm.init();
-        }).then(() => {
-            logger.log("Using legacy Olm");
-        }).catch((e) => {
-            logger.log("Both WebAssembly and asm.js Olm failed!", e);
-        });
-    });
-}
-
-export async function loadLanguage() {
-    const prefLang = SettingsStore.getValue("language", null, /*excludeDefault=*/true);
-    let langs = [];
+export async function loadLanguage(): Promise<void> {
+    const prefLang = SettingsStore.getValue("language", null, /*excludeDefault=*/ true);
+    let langs: string[] = [];
 
     if (!prefLang) {
         languageHandler.getLanguagesFromBrowser().forEach((l) => {
@@ -127,34 +82,79 @@ export async function loadLanguage() {
     }
 }
 
-export async function loadTheme() {
-    setTheme();
+export async function loadTheme(): Promise<void> {
+    return setTheme();
 }
 
-export async function loadApp(fragParams: {}) {
+export async function loadApp(fragParams: QueryDict): Promise<void> {
     // load app.js async so that its code is not executed immediately and we can catch any exceptions
     const module = await import(
         /* webpackChunkName: "element-web-app" */
         /* webpackPreload: true */
-        "./app");
-    window.matrixChat = ReactDOM.render(await module.loadApp(fragParams),
-        document.getElementById('matrixchat'));
+        "./app"
+    );
+    function setWindowMatrixChat(matrixChat: MatrixChat): void {
+        window.matrixChat = matrixChat;
+    }
+    const app = await module.loadApp(fragParams, setWindowMatrixChat);
+    const root = createRoot(document.getElementById("matrixchat")!);
+    root.render(app);
 }
 
-export async function showError(title: string, messages?: string[]) {
-    const ErrorView = (await import(
+export async function showError(title: string, messages?: string[]): Promise<void> {
+    const { ErrorView } = await import(
         /* webpackChunkName: "error-view" */
-        "../async-components/structures/ErrorView")).default;
-    window.matrixChat = ReactDOM.render(<ErrorView title={title} messages={messages} />,
-        document.getElementById('matrixchat'));
+        "../async-components/structures/ErrorView"
+    );
+    const root = createRoot(document.getElementById("matrixchat")!);
+    root.render(
+        <StrictMode>
+            <ErrorView title={title} messages={messages} />
+        </StrictMode>,
+    );
 }
 
-export async function showIncompatibleBrowser(onAccept) {
-    const CompatibilityView = (await import(
-        /* webpackChunkName: "compatibility-view" */
-        "../async-components/structures/CompatibilityView")).default;
-    window.matrixChat = ReactDOM.render(<CompatibilityView onAccept={onAccept} />,
-        document.getElementById('matrixchat'));
+export async function showIncompatibleBrowser(onAccept: () => void): Promise<void> {
+    const { UnsupportedBrowserView } = await import(
+        /* webpackChunkName: "error-view" */
+        "../async-components/structures/ErrorView"
+    );
+    const root = createRoot(document.getElementById("matrixchat")!);
+    root.render(
+        <StrictMode>
+            <UnsupportedBrowserView onAccept={onAccept} />
+        </StrictMode>,
+    );
 }
 
-export const _t = languageHandler._t;
+/**
+ * @deprecated in favour of the plugin system
+ */
+export async function loadModules(): Promise<void> {
+    const { INSTALLED_MODULES } = await import("../modules.js");
+    for (const InstalledModule of INSTALLED_MODULES) {
+        ModuleRunner.instance.registerModule((api) => new InstalledModule(api));
+    }
+}
+
+export async function loadPlugins(): Promise<void> {
+    // Add React to the global namespace, this is part of the new Module API contract to avoid needing
+    // every single module to ship its own copy of React. This also makes it easier to access via the console
+    // and incidentally means we can forget our React imports in JSX files without penalty.
+    window.React = React;
+
+    const modules = SdkConfig.get("modules");
+    if (!modules?.length) return;
+    const moduleLoader = new ModuleLoader(ModuleApi);
+    window.mxModuleLoader = moduleLoader;
+    for (const src of modules) {
+        // We need to instruct webpack to not mangle this import as it is not available at compile time
+        const module = await import(/* webpackIgnore: true */ src);
+        await moduleLoader.load(module);
+    }
+    await moduleLoader.start();
+}
+
+export { _t } from "../languageHandler";
+
+export { extractErrorMessageFromError } from "../components/views/dialogs/ErrorDialog";
