@@ -20,7 +20,7 @@
 
 class Optional {
     static from(value) {
-        return value && Some.of(value) || None;
+        return (value && Some.of(value)) || None;
     }
     map(f) {
         return this;
