@@ -1,108 +1,66 @@
 /*
-Copyright 2015, 2016 OpenMarket Ltd
-Copyright 2017 Vector Creations Ltd
-Copyright 2018, 2019 New Vector Ltd
-Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
+Copyright 2024 New Vector Ltd.
 Copyright 2020 The Matrix.org Foundation C.I.C.
+Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
+Copyright 2018, 2019 New Vector Ltd
+Copyright 2017 Vector Creations Ltd
+Copyright 2015, 2016 OpenMarket Ltd
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
 */
 
-// To ensure we load the browser-request version
-import "matrix-js-sdk"; // eslint-disable-line no-restricted-imports
-
-import React from 'react';
-import PlatformPeg from 'matrix-react-sdk/src/PlatformPeg';
-import { _td, newTranslatableError } from 'matrix-react-sdk/src/languageHandler';
-import AutoDiscoveryUtils from 'matrix-react-sdk/src/utils/AutoDiscoveryUtils';
-import { AutoDiscovery } from "matrix-js-sdk/src/autodiscovery";
-import * as Lifecycle from "matrix-react-sdk/src/Lifecycle";
-import SdkConfig, { parseSsoRedirectOptions } from "matrix-react-sdk/src/SdkConfig";
-import { IConfigOptions } from "matrix-react-sdk/src/IConfigOptions";
+// To ensure we load the browser-matrix version first
+import "matrix-js-sdk/src/browser-index";
+import React, { type ReactElement, StrictMode } from "react";
 import { logger } from "matrix-js-sdk/src/logger";
-import { createClient } from "matrix-js-sdk/src/matrix";
-import { SnakedObject } from "matrix-react-sdk/src/utils/SnakedObject";
-import MatrixChat from "matrix-react-sdk/src/components/structures/MatrixChat";
+import { createClient, AutoDiscovery, type ClientConfig } from "matrix-js-sdk/src/matrix";
+import { WrapperLifecycle, type WrapperOpts } from "@matrix-org/react-sdk-module-api/lib/lifecycles/WrapperLifecycle";
 
-import { parseQs } from './url_utils';
-import VectorBasePlatform from "./platform/VectorBasePlatform";
-import { getScreenFromLocation, init as initRouting, onNewScreen } from "./routing";
-
-// add React and ReactPerf to the global namespace, to make them easier to access via the console
-// this incidentally means we can forget our React imports in JSX files without penalty.
-window.React = React;
+import type { QueryDict } from "matrix-js-sdk/src/utils";
+import PlatformPeg from "../PlatformPeg";
+import AutoDiscoveryUtils from "../utils/AutoDiscoveryUtils";
+import * as Lifecycle from "../Lifecycle";
+import SdkConfig, { parseSsoRedirectOptions } from "../SdkConfig";
+import { type IConfigOptions } from "../IConfigOptions";
+import { SnakedObject } from "../utils/SnakedObject";
+import MatrixChat from "../components/structures/MatrixChat";
+import { type ValidatedServerConfig } from "../utils/ValidatedServerConfig";
+import { ModuleRunner } from "../modules/ModuleRunner";
+import { parseQs } from "./url_utils";
+import { getInitialScreenAfterLogin, getScreenFromLocation, init as initRouting, onNewScreen } from "./routing";
+import { UserFriendlyError } from "../languageHandler";
 
 logger.log(`Application is running in ${process.env.NODE_ENV} mode`);
 
 window.matrixLogger = logger;
 
-// We use this to work out what URL the SDK should
-// pass through when registering to allow the user to
-// click back to the client having registered.
-// It's up to us to recognise if we're loaded with
-// this URL and tell MatrixClient to resume registration.
-//
-// If we're in electron, we should never pass through a file:// URL otherwise
-// the identity server will try to 302 the browser to it, which breaks horribly.
-// so in that instance, hardcode to use app.element.io for now instead.
-function makeRegistrationUrl(params: object) {
-    let url;
-    if (window.location.protocol === "vector:") {
-        url = 'https://app.element.io/#/register';
-    } else {
-        url = (
-            window.location.protocol + '//' +
-            window.location.host +
-            window.location.pathname +
-            '#/register'
-        );
-    }
-
-    const keys = Object.keys(params);
-    for (let i = 0; i < keys.length; ++i) {
-        if (i === 0) {
-            url += '?';
-        } else {
-            url += '&';
-        }
-        const k = keys[i];
-        url += k + '=' + encodeURIComponent(params[k]);
-    }
-    return url;
-}
-
-function onTokenLoginCompleted() {
+function onTokenLoginCompleted(): void {
     // if we did a token login, we're now left with the token, hs and is
-    // url as query params in the url; a little nasty but let's redirect to
-    // clear them.
+    // url as query params in the url;
+    // if we did an oidc authorization code flow login, we're left with the auth code and state
+    // as query params in the url;
+    // a little nasty but let's redirect to clear them.
     const url = new URL(window.location.href);
 
     url.searchParams.delete("loginToken");
+    url.searchParams.delete("state");
+    url.searchParams.delete("code");
 
-    logger.log(`Redirecting to ${url.href} to drop loginToken from queryparams`);
+    logger.log(`Redirecting to ${url.href} to drop delegated authentication params from queryparams`);
     window.history.replaceState(null, "", url.href);
 }
 
-export async function loadApp(fragParams: {}) {
+export async function loadApp(fragParams: QueryDict, matrixChatRef: React.Ref<MatrixChat>): Promise<ReactElement> {
     initRouting();
     const platform = PlatformPeg.get();
 
     const params = parseQs(window.location);
 
-    const urlWithoutQuery = window.location.protocol + '//' + window.location.host + window.location.pathname;
+    const urlWithoutQuery = window.location.protocol + "//" + window.location.host + window.location.pathname;
     logger.log("Vector starting at " + urlWithoutQuery);
 
-    (platform as VectorBasePlatform).startUpdater();
+    platform?.startUpdater();
 
     // Don't bother loading the app until the config is verified
     const config = await verifyServerConfig();
@@ -115,42 +73,59 @@ export async function loadApp(fragParams: {}) {
     const ssoRedirects = parseSsoRedirectOptions(config);
     let autoRedirect = ssoRedirects.immediate === true;
     // XXX: This path matching is a bit brittle, but better to do it early instead of in the app code.
-    const isWelcomeOrLanding = window.location.hash === '#/welcome' || window.location.hash === '#';
+    const isWelcomeOrLanding =
+        window.location.hash === "#/welcome" || window.location.hash === "#" || window.location.hash === "";
+    const isLoginPage = window.location.hash === "#/login";
+
     if (!autoRedirect && ssoRedirects.on_welcome_page && isWelcomeOrLanding) {
+        autoRedirect = true;
+    }
+    if (!autoRedirect && ssoRedirects.on_login_page && isLoginPage) {
         autoRedirect = true;
     }
     if (!hasPossibleToken && !isReturningFromSso && autoRedirect) {
         logger.log("Bypassing app load to redirect to SSO");
         const tempCli = createClient({
-            baseUrl: config.validated_server_config.hsUrl,
-            idBaseUrl: config.validated_server_config.isUrl,
+            baseUrl: config.validated_server_config!.hsUrl,
+            idBaseUrl: config.validated_server_config!.isUrl,
         });
-        PlatformPeg.get().startSingleSignOn(tempCli, "sso", `/${getScreenFromLocation(window.location).screen}`);
+        PlatformPeg.get()!.startSingleSignOn(tempCli, "sso", `/${getScreenFromLocation(window.location).screen}`);
 
         // We return here because startSingleSignOn() will asynchronously redirect us. We don't
         // care to wait for it, and don't want to show any UI while we wait (not even half a welcome
         // page). As such, just don't even bother loading the MatrixChat component.
-        return;
+        return <React.Fragment />;
     }
 
-    const defaultDeviceName = snakedConfig.get("default_device_display_name")
-        ?? platform.getDefaultDeviceDisplayName();
+    const defaultDeviceName =
+        snakedConfig.get("default_device_display_name") ?? platform?.getDefaultDeviceDisplayName();
 
-    return <MatrixChat
-        onNewScreen={onNewScreen}
-        makeRegistrationUrl={makeRegistrationUrl}
-        config={config}
-        realQueryParams={params}
-        startingFragmentQueryParams={fragParams}
-        enableGuest={!config.disable_guests}
-        onTokenLoginCompleted={onTokenLoginCompleted}
-        initialScreenAfterLogin={getScreenFromLocation(window.location)}
-        defaultDeviceDisplayName={defaultDeviceName}
-    />;
+    const initialScreenAfterLogin = getInitialScreenAfterLogin(window.location);
+
+    const wrapperOpts: WrapperOpts = { Wrapper: React.Fragment };
+    ModuleRunner.instance.invoke(WrapperLifecycle.Wrapper, wrapperOpts);
+
+    return (
+        <wrapperOpts.Wrapper>
+            <StrictMode>
+                <MatrixChat
+                    ref={matrixChatRef}
+                    onNewScreen={onNewScreen}
+                    config={config}
+                    realQueryParams={params}
+                    startingFragmentQueryParams={fragParams}
+                    enableGuest={!config.disable_guests}
+                    onTokenLoginCompleted={onTokenLoginCompleted}
+                    initialScreenAfterLogin={initialScreenAfterLogin}
+                    defaultDeviceDisplayName={defaultDeviceName}
+                />
+            </StrictMode>
+        </wrapperOpts.Wrapper>
+    );
 }
 
 async function verifyServerConfig(): Promise<IConfigOptions> {
-    let validatedConfig;
+    let validatedConfig: ValidatedServerConfig;
     try {
         logger.log("Verifying homeserver configuration");
 
@@ -164,45 +139,42 @@ async function verifyServerConfig(): Promise<IConfigOptions> {
         // validators for that purpose.
 
         const config = SdkConfig.get();
-        let wkConfig = config['default_server_config']; // overwritten later under some conditions
-        const serverName = config['default_server_name'];
-        const hsUrl = config['default_hs_url'];
-        const isUrl = config['default_is_url'];
+        let wkConfig = config["default_server_config"]; // overwritten later under some conditions
+        const serverName = config["default_server_name"];
+        const hsUrl = config["default_hs_url"];
+        const isUrl = config["default_is_url"];
 
-        const incompatibleOptions = [wkConfig, serverName, hsUrl].filter(i => !!i);
-        if (incompatibleOptions.length > 1) {
+        const incompatibleOptions = [wkConfig, serverName, hsUrl].filter((i) => !!i);
+        if (hsUrl && (wkConfig || serverName)) {
             // noinspection ExceptionCaughtLocallyJS
-            throw newTranslatableError(_td(
-                "Invalid configuration: can only specify one of default_server_config, default_server_name, " +
-                "or default_hs_url.",
-            ));
+            throw new UserFriendlyError("error|invalid_configuration_mixed_server");
         }
         if (incompatibleOptions.length < 1) {
             // noinspection ExceptionCaughtLocallyJS
-            throw newTranslatableError(_td("Invalid configuration: no default server specified."));
+            throw new UserFriendlyError("error|invalid_configuration_no_server");
         }
 
         if (hsUrl) {
             logger.log("Config uses a default_hs_url - constructing a default_server_config using this information");
             logger.warn(
                 "DEPRECATED CONFIG OPTION: In the future, default_hs_url will not be accepted. Please use " +
-                "default_server_config instead.",
+                    "default_server_config instead.",
             );
 
             wkConfig = {
                 "m.homeserver": {
-                    "base_url": hsUrl,
+                    base_url: hsUrl,
                 },
             };
             if (isUrl) {
                 wkConfig["m.identity_server"] = {
-                    "base_url": isUrl,
+                    base_url: isUrl,
                 };
             }
         }
 
-        let discoveryResult = null;
-        if (wkConfig) {
+        let discoveryResult: ClientConfig | undefined;
+        if (!serverName && wkConfig) {
             logger.log("Config uses a default_server_config - validating object");
             discoveryResult = await AutoDiscovery.fromDiscoveryConfig(wkConfig);
         }
@@ -211,12 +183,16 @@ async function verifyServerConfig(): Promise<IConfigOptions> {
             logger.log("Config uses a default_server_name - doing .well-known lookup");
             logger.warn(
                 "DEPRECATED CONFIG OPTION: In the future, default_server_name will not be accepted. Please " +
-                "use default_server_config instead.",
+                    "use default_server_config instead.",
             );
             discoveryResult = await AutoDiscovery.findClientConfig(serverName);
+            if (discoveryResult["m.homeserver"].base_url === null && wkConfig) {
+                logger.log("Finding base_url failed but a default_server_config was found - using it as a fallback");
+                discoveryResult = await AutoDiscovery.fromDiscoveryConfig(wkConfig);
+            }
         }
 
-        validatedConfig = AutoDiscoveryUtils.buildValidatedConfigFromDiscovery(serverName, discoveryResult, true);
+        validatedConfig = await AutoDiscoveryUtils.buildValidatedConfigFromDiscovery(serverName, discoveryResult, true);
     } catch (e) {
         const { hsUrl, isUrl, userId } = await Lifecycle.getStoredSessionVars();
         if (hsUrl && userId) {
@@ -238,7 +214,7 @@ async function verifyServerConfig(): Promise<IConfigOptions> {
 
     // Add the newly built config to the actual config for use by the app
     logger.log("Updating SdkConfig with validated discovery information");
-    SdkConfig.add({ "validated_server_config": validatedConfig });
+    SdkConfig.add({ validated_server_config: validatedConfig });
 
     return SdkConfig.get();
 }
