@@ -1,42 +1,76 @@
 /*
-Copyright 2020 New Vector Ltd.
+Copyright 2020-2024 New Vector Ltd.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
 */
 
-import { KJUR } from 'jsrsasign';
+import { KJUR } from "jsrsasign";
 import {
-    IOpenIDCredentials,
-    IWidgetApiRequest,
+    type IOpenIDCredentials,
+    type IWidgetApiRequest,
+    type IWidgetApiRequestData,
+    type IWidgetApiResponseData,
     VideoConferenceCapabilities,
     WidgetApi,
+    type WidgetApiAction,
 } from "matrix-widget-api";
-import { ElementWidgetActions } from "matrix-react-sdk/src/stores/widgets/ElementWidgetActions";
 import { logger } from "matrix-js-sdk/src/logger";
-import { IConfigOptions } from "matrix-react-sdk/src/IConfigOptions";
-import { SnakedObject } from "matrix-react-sdk/src/utils/SnakedObject";
 
+import type {
+    JitsiMeetExternalAPIConstructor,
+    ExternalAPIEventCallbacks,
+    JitsiMeetExternalAPI as _JitsiMeetExternalAPI,
+    LogEvent,
+    VideoMuteStatusChangedEvent,
+    ExternalAPIOptions as _ExternalAPIOptions,
+    Config as _Config,
+    InterfaceConfig as _InterfaceConfig,
+} from "jitsi-meet";
+import { ElementWidgetActions } from "../../stores/widgets/ElementWidgetActions";
+import { type IConfigOptions } from "../../IConfigOptions";
+import { SnakedObject } from "../../utils/SnakedObject";
+import { ElementWidgetCapabilities } from "../../stores/widgets/ElementWidgetCapabilities";
 import { getVectorConfig } from "../getconfig";
 
-// We have to trick webpack into loading our CSS for us.
-require("./index.scss");
+interface Config extends _Config {
+    // Jitsi's types are missing these fields
+    prejoinConfig?: {
+        enabled: boolean;
+        hideDisplayName?: boolean;
+        hideExtraJoinButtons?: string[];
+    };
+    toolbarButtons?: string[];
+    conferenceInfo?: {
+        alwaysVIsible?: string[];
+        autoHide?: string[];
+    };
+    disableSelfViewSettings?: boolean;
+}
 
-const JITSI_OPENIDTOKEN_JWT_AUTH = 'openidtoken-jwt';
+interface InterfaceConfig extends _InterfaceConfig {
+    // XXX: It is unclear whether this is a typo of TOOLBAR_BUTTONS or if its just really undocumented,
+    // either way it is missing in types, yet we try and use it
+    MAIN_TOOLBAR_BUTTONS?: string[];
+}
+
+interface ExternalAPIOptions extends _ExternalAPIOptions {
+    configOverwrite?: Config;
+    interfaceConfigOverwrite?: InterfaceConfig;
+    // Jitsi's types are missing these fields
+    lang?: string;
+}
+
+// We have to trick webpack into loading our CSS for us.
+// eslint-disable-next-line @typescript-eslint/no-require-imports
+require("./index.pcss");
+
+const JITSI_OPENIDTOKEN_JWT_AUTH = "openidtoken-jwt";
 
 // Dev note: we use raw JS without many dependencies to reduce bundle size.
 // We do not need all of React to render a Jitsi conference.
 
-declare let JitsiMeetExternalAPI: any;
+declare let JitsiMeetExternalAPI: JitsiMeetExternalAPIConstructor;
 
 let inConference = false;
 
@@ -48,22 +82,31 @@ let avatarUrl: string;
 let userId: string;
 let jitsiAuth: string;
 let roomId: string;
-let openIdToken: IOpenIDCredentials;
 let roomName: string;
 let startAudioOnly: boolean;
+let startWithAudioMuted: boolean | undefined;
+let startWithVideoMuted: boolean | undefined;
 let isVideoChannel: boolean;
+let supportsScreensharing: boolean;
+let language: string;
 
-let widgetApi: WidgetApi;
-let meetApi: any; // JitsiMeetExternalAPI
+let widgetApi: WidgetApi | undefined;
+let meetApi: _JitsiMeetExternalAPI | undefined;
 let skipOurWelcomeScreen = false;
 
-const ack = (ev: CustomEvent<IWidgetApiRequest>) => widgetApi.transport.reply(ev.detail, {});
+async function checkAudioVideoEnabled(): Promise<[audioEnabled: boolean, videoEnabled: boolean]> {
+    if (!meetApi) return [false, false];
+    const [audioEnabled, videoEnabled] = (await Promise.all([meetApi.isAudioMuted(), meetApi.isVideoMuted()])).map(
+        (muted) => !muted,
+    );
+    return [audioEnabled, videoEnabled];
+}
 
-(async function() {
+const setupCompleted = (async (): Promise<string | void> => {
     try {
         // Queue a config.json lookup asap, so we can use it later on. We want this to be concurrent with
         // other setup work and therefore do not block.
-        const configPromise = getVectorConfig('..');
+        const configPromise = getVectorConfig();
 
         // The widget's options are encoded into the fragment to avoid leaking info to the server.
         const widgetQuery = new URLSearchParams(window.location.hash.substring(1));
@@ -76,143 +119,148 @@ const ack = (ev: CustomEvent<IWidgetApiRequest>) => widgetApi.transport.reply(ev
             }
             return vals[0];
         };
+        const parseBooleanOrUndefined = (value: string | undefined): boolean | undefined => {
+            if (value === undefined) {
+                return undefined;
+            }
+
+            return value === "true";
+        };
 
         // If we have these params, expect a widget API to be available (ie. to be in an iframe
         // inside a matrix client). Otherwise, assume we're on our own, eg. have been popped
         // out into a browser.
-        const parentUrl = qsParam('parentUrl', true);
-        const widgetId = qsParam('widgetId', true);
-        const theme = qsParam('theme', true);
+        const parentUrl = qsParam("parentUrl", true);
+        const widgetId = qsParam("widgetId", true);
+        const theme = qsParam("theme", true);
+        language = qsParam("language", true) ?? "en";
 
         if (theme) {
             document.body.classList.add(`theme-${theme.replace(" ", "_")}`);
         }
 
         // Set this up as early as possible because Element will be hitting it almost immediately.
-        let readyPromise: Promise<[void, void]>;
+        let widgetApiReady: Promise<void> | undefined;
         if (parentUrl && widgetId) {
-            const parentOrigin = new URL(qsParam('parentUrl')).origin;
+            const parentOrigin = new URL(qsParam("parentUrl")).origin;
             widgetApi = new WidgetApi(qsParam("widgetId"), parentOrigin);
+
+            widgetApiReady = new Promise<void>((resolve) => widgetApi!.once("ready", resolve));
             widgetApi.requestCapabilities(VideoConferenceCapabilities);
-            readyPromise = Promise.all([
-                new Promise<void>(resolve => {
-                    widgetApi.once(`action:${ElementWidgetActions.ClientReady}`, ev => {
-                        ev.preventDefault();
-                        widgetApi.transport.reply(ev.detail, {});
-                        resolve();
-                    });
-                }),
-                new Promise<void>(resolve => {
-                    widgetApi.once("ready", () => resolve());
-                }),
-            ]);
+
+            // jitsi cannot work in a popup if auth token is provided because widgetApi is not available there
+            // so check the token and request the 'requires_client' capability to hide the popup icon in the Element
+            if (qsParam("auth", true) === "openidtoken-jwt") {
+                widgetApi.requestCapability(ElementWidgetCapabilities.RequiresClient);
+            }
+
             widgetApi.start();
+
+            const handleAction = (
+                action: WidgetApiAction,
+                handler: (request: IWidgetApiRequestData) => Promise<IWidgetApiResponseData | void>,
+            ): void => {
+                widgetApi!.on(`action:${action}`, async (ev: CustomEvent<IWidgetApiRequest>) => {
+                    ev.preventDefault();
+                    await setupCompleted;
+
+                    let response: IWidgetApiResponseData;
+                    try {
+                        response = (await handler(ev.detail.data)) ?? {};
+                    } catch (e) {
+                        if (e instanceof Error) {
+                            response = { error: { message: e.message } };
+                        } else {
+                            throw e;
+                        }
+                    }
+
+                    widgetApi!.transport.reply(ev.detail, response);
+                });
+            };
+
+            handleAction(ElementWidgetActions.JoinCall, async ({ audioInput, videoInput }) => {
+                void joinConference(audioInput as string | null, videoInput as string | null);
+            });
+            handleAction(ElementWidgetActions.HangupCall, async ({ force }) => {
+                if (force === true) {
+                    meetApi?.dispose();
+                    void notifyHangup();
+                    meetApi = undefined;
+                    closeConference();
+                } else {
+                    meetApi?.executeCommand("hangup");
+                }
+            });
+            handleAction(ElementWidgetActions.DeviceMute, async (params) => {
+                if (!meetApi) return;
+
+                const [audioEnabled, videoEnabled] = await checkAudioVideoEnabled();
+
+                if (Object.keys(params).length === 0) {
+                    // Handle query
+                    return {
+                        audio_enabled: audioEnabled,
+                        video_enabled: videoEnabled,
+                    };
+                }
+
+                if (params.audio_enabled !== audioEnabled) {
+                    meetApi.executeCommand("toggleAudio");
+                }
+                if (params.video_enabled !== videoEnabled) {
+                    meetApi.executeCommand("toggleVideo");
+                }
+                return params;
+            });
+            handleAction(ElementWidgetActions.TileLayout, async () => {
+                meetApi?.executeCommand("setTileView", true);
+            });
+            handleAction(ElementWidgetActions.SpotlightLayout, async () => {
+                meetApi?.executeCommand("setTileView", false);
+            });
+            handleAction(ElementWidgetActions.StartLiveStream, async ({ rtmpStreamKey }) => {
+                if (!meetApi) throw new Error("Conference not joined");
+                meetApi.executeCommand("startRecording", {
+                    mode: "stream",
+                    // this looks like it should be rtmpStreamKey but we may be on too old
+                    // a version of jitsi meet
+                    //rtmpStreamKey,
+                    youtubeStreamKey: rtmpStreamKey,
+                });
+            });
         } else {
             logger.warn("No parent URL or no widget ID - assuming no widget API is available");
         }
 
         // Populate the Jitsi params now
-        jitsiDomain = qsParam('conferenceDomain');
-        conferenceId = qsParam('conferenceId');
-        displayName = qsParam('displayName', true);
-        avatarUrl = qsParam('avatarUrl', true); // http not mxc
-        userId = qsParam('userId');
-        jitsiAuth = qsParam('auth', true);
-        roomId = qsParam('roomId', true);
-        roomName = qsParam('roomName', true);
-        startAudioOnly = qsParam('isAudioOnly', true) === "true";
-        isVideoChannel = qsParam('isVideoChannel', true) === "true";
+        jitsiDomain = qsParam("conferenceDomain");
+        conferenceId = qsParam("conferenceId");
+        displayName = qsParam("displayName", true);
+        avatarUrl = qsParam("avatarUrl", true); // http not mxc
+        userId = qsParam("userId");
+        jitsiAuth = qsParam("auth", true);
+        roomId = qsParam("roomId", true);
+        roomName = qsParam("roomName", true);
+        startAudioOnly = qsParam("isAudioOnly", true) === "true";
+        startWithAudioMuted = parseBooleanOrUndefined(qsParam("startWithAudioMuted", true));
+        startWithVideoMuted = parseBooleanOrUndefined(qsParam("startWithVideoMuted", true));
+        isVideoChannel = qsParam("isVideoChannel", true) === "true";
+        supportsScreensharing = qsParam("supportsScreensharing", true) === "true";
 
         // We've reached the point where we have to wait for the config, so do that then parse it.
         const instanceConfig = new SnakedObject<IConfigOptions>((await configPromise) ?? <IConfigOptions>{});
-        const jitsiConfig = instanceConfig.get("jitsi_widget") ?? {};
-        skipOurWelcomeScreen = (new SnakedObject<IConfigOptions["jitsi_widget"]>(jitsiConfig))
-            .get("skip_built_in_welcome_screen") ?? false;
+        const jitsiConfig = instanceConfig.get("jitsi_widget");
+        if (jitsiConfig) {
+            skipOurWelcomeScreen = new SnakedObject(jitsiConfig).get("skip_built_in_welcome_screen") ?? false;
+        }
 
         // Either reveal the prejoin screen, or skip straight to Jitsi depending on the config.
         // We don't set up the call yet though as this might lead to failure without the widget API.
         toggleConferenceVisibility(skipOurWelcomeScreen);
 
         if (widgetApi) {
-            await readyPromise;
-
-            // See https://github.com/matrix-org/prosody-mod-auth-matrix-user-verification
-            if (jitsiAuth === JITSI_OPENIDTOKEN_JWT_AUTH) {
-                // Request credentials, give callback to continue when received
-                openIdToken = await widgetApi.requestOpenIDConnectToken();
-                logger.log("Got OpenID Connect token");
-            }
-
-            widgetApi.on(`action:${ElementWidgetActions.JoinCall}`,
-                (ev: CustomEvent<IWidgetApiRequest>) => {
-                    const { audioDevice, videoDevice } = ev.detail.data;
-                    joinConference(audioDevice as string, videoDevice as string);
-                    ack(ev);
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.HangupCall}`,
-                (ev: CustomEvent<IWidgetApiRequest>) => {
-                    meetApi?.executeCommand('hangup');
-                    ack(ev);
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.ForceHangupCall}`,
-                (ev: CustomEvent<IWidgetApiRequest>) => {
-                    meetApi?.dispose();
-                    notifyHangup();
-                    meetApi = null;
-                    closeConference();
-                    ack(ev);
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.MuteAudio}`,
-                async (ev: CustomEvent<IWidgetApiRequest>) => {
-                    ack(ev);
-                    if (meetApi && !await meetApi.isAudioMuted()) {
-                        meetApi.executeCommand('toggleAudio');
-                    }
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.UnmuteAudio}`,
-                async (ev: CustomEvent<IWidgetApiRequest>) => {
-                    ack(ev);
-                    if (meetApi && await meetApi.isAudioMuted()) {
-                        meetApi.executeCommand('toggleAudio');
-                    }
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.MuteVideo}`,
-                async (ev: CustomEvent<IWidgetApiRequest>) => {
-                    ack(ev);
-                    if (meetApi && !await meetApi.isVideoMuted()) {
-                        meetApi.executeCommand('toggleVideo');
-                    }
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.UnmuteVideo}`,
-                async (ev: CustomEvent<IWidgetApiRequest>) => {
-                    ack(ev);
-                    if (meetApi && await meetApi.isVideoMuted()) {
-                        meetApi.executeCommand('toggleVideo');
-                    }
-                },
-            );
-            widgetApi.on(`action:${ElementWidgetActions.StartLiveStream}`,
-                (ev: CustomEvent<IWidgetApiRequest>) => {
-                    if (meetApi) {
-                        meetApi.executeCommand('startRecording', {
-                            mode: 'stream',
-                            // this looks like it should be rtmpStreamKey but we may be on too old
-                            // a version of jitsi meet
-                            //rtmpStreamKey: ev.detail.data.rtmpStreamKey,
-                            youtubeStreamKey: ev.detail.data.rtmpStreamKey,
-                        });
-                        ack(ev);
-                    } else {
-                        widgetApi.transport.reply(ev.detail, { error: { message: "Conference not joined" } });
-                    }
-                },
-            );
+            await widgetApiReady;
         }
 
         // Now that everything should be set up, skip to the Jitsi splash screen if needed
@@ -221,24 +269,17 @@ const ack = (ev: CustomEvent<IWidgetApiRequest>) => widgetApi.transport.reply(ev
         }
 
         enableJoinButton(); // always enable the button
-
-        // Inform the client that we're ready to receive events
-        try {
-            await widgetApi?.transport.send(ElementWidgetActions.WidgetReady, {});
-        } catch (e) {
-            logger.error(e);
-        }
     } catch (e) {
         logger.error("Error setting up Jitsi widget", e);
-        document.getElementById("widgetActionContainer").innerText = "Failed to load Jitsi widget";
+        document.getElementById("widgetActionContainer")!.innerText = "Failed to load Jitsi widget";
     }
 })();
 
-function enableJoinButton() {
-    document.getElementById("joinButton").onclick = () => joinConference();
+function enableJoinButton(): void {
+    document.getElementById("joinButton")!.onclick = (): Promise<void> => joinConference();
 }
 
-function switchVisibleContainers() {
+function switchVisibleContainers(): void {
     inConference = !inConference;
 
     // Our welcome screen is managed by other code, so just don't switch to it ever
@@ -248,26 +289,24 @@ function switchVisibleContainers() {
     }
 }
 
-function toggleConferenceVisibility(inConference: boolean) {
-    document.getElementById("jitsiContainer").style.visibility = inConference ? 'unset' : 'hidden';
-    // Video rooms have a separate UI for joining, so they should never show our join button
-    document.getElementById("joinButtonContainer").style.visibility =
-        (inConference || isVideoChannel) ? 'hidden' : 'unset';
+function toggleConferenceVisibility(inConference: boolean): void {
+    document.getElementById("jitsiContainer")!.style.visibility = inConference ? "unset" : "hidden";
+    document.getElementById("joinButtonContainer")!.style.visibility = inConference ? "hidden" : "unset";
 }
 
-function skipToJitsiSplashScreen() {
+function skipToJitsiSplashScreen(): void {
     // really just a function alias for self-documenting code
-    joinConference();
+    void joinConference();
 }
 
 /**
- * Create a JWT token fot jitsi openidtoken-jwt auth
+ * Create a JWT token for jitsi openidtoken-jwt auth
  *
  * See https://github.com/matrix-org/prosody-mod-auth-matrix-user-verification
  */
-function createJWTToken() {
+function createJWTToken(openIdToken: IOpenIDCredentials): string {
     // Header
-    const header = { alg: 'HS256', typ: 'JWT' };
+    const header = { alg: "HS256", typ: "JWT" };
     // Payload
     const payload = {
         // As per Jitsi token auth, `iss` needs to be set to something agreed between
@@ -292,15 +331,10 @@ function createJWTToken() {
     // Sign JWT
     // The secret string here is irrelevant, we're only using the JWT
     // to transport data to Prosody in the Jitsi stack.
-    return KJUR.jws.JWS.sign(
-        'HS256',
-        JSON.stringify(header),
-        JSON.stringify(payload),
-        'notused',
-    );
+    return KJUR.jws.JWS.sign("HS256", JSON.stringify(header), JSON.stringify(payload), "notused");
 }
 
-async function notifyHangup(errorMessage?: string) {
+async function notifyHangup(errorMessage?: string): Promise<void> {
     if (widgetApi) {
         // We send the hangup event before setAlwaysOnScreen, because the latter
         // can cause the receiving side to instantly stop listening.
@@ -312,44 +346,78 @@ async function notifyHangup(errorMessage?: string) {
     }
 }
 
-function closeConference() {
+function closeConference(): void {
     switchVisibleContainers();
-    document.getElementById("jitsiContainer").innerHTML = "";
+    document.getElementById("jitsiContainer")!.innerHTML = "";
 
     if (skipOurWelcomeScreen) {
         skipToJitsiSplashScreen();
     }
 }
 
+// Converts from IETF language tags used by Element (en-US) to the format used
+// by Jitsi (enUS)
+function normalizeLanguage(language: string): string {
+    const [lang, variant] = language.replace("_", "-").split("-");
+
+    if (!variant || lang === variant) {
+        return lang;
+    }
+
+    return lang + variant.toUpperCase();
+}
+
+function mapLanguage(language: string): string {
+    // Element and Jitsi don't agree how to interpret en, so we go with Elements
+    // interpretation to stay consistent
+    switch (language) {
+        case "en":
+            return "enGB";
+        case "enUS":
+            return "en";
+        default:
+            return language;
+    }
+}
+
 // event handler bound in HTML
-function joinConference(audioDevice?: string, videoDevice?: string) {
-    let jwt;
+// An audio input of undefined instructs Jitsi to start unmuted with whatever
+// audio input it can find, while an input of null instructs it to start muted,
+// and a non-nullish input specifies the label of a specific device to use.
+// Same for video inputs.
+async function joinConference(audioInput?: string | null, videoInput?: string | null): Promise<void> {
+    let jwt: string | undefined;
     if (jitsiAuth === JITSI_OPENIDTOKEN_JWT_AUTH) {
-        if (!openIdToken?.access_token) { // eslint-disable-line camelcase
+        // See https://github.com/matrix-org/prosody-mod-auth-matrix-user-verification
+        const openIdToken = await widgetApi?.requestOpenIDConnectToken();
+        logger.log("Got OpenID Connect token");
+
+        if (!openIdToken?.access_token) {
+            // eslint-disable-line camelcase
             // We've failing to get a token, don't try to init conference
-            logger.warn('Expected to have an OpenID credential, cannot initialize widget.');
-            document.getElementById("widgetActionContainer").innerText = "Failed to load Jitsi widget";
+            logger.warn("Expected to have an OpenID credential, cannot initialize widget.");
+            document.getElementById("widgetActionContainer")!.innerText = "Failed to load Jitsi widget";
             return;
         }
-        jwt = createJWTToken();
+        jwt = createJWTToken(openIdToken);
     }
 
     switchVisibleContainers();
 
     logger.warn(
         "[Jitsi Widget] The next few errors about failing to parse URL parameters are fine if " +
-        "they mention 'external_api' or 'jitsi' in the stack. They're just Jitsi Meet trying to parse " +
-        "our fragment values and not recognizing the options.",
+            "they mention 'external_api' or 'jitsi' in the stack. They're just Jitsi Meet trying to parse " +
+            "our fragment values and not recognizing the options.",
     );
 
-    const options = {
+    const options: ExternalAPIOptions = {
         width: "100%",
         height: "100%",
-        parentNode: document.querySelector("#jitsiContainer"),
+        parentNode: document.querySelector("#jitsiContainer") ?? undefined,
         roomName: conferenceId,
         devices: {
-            audioInput: audioDevice,
-            videoInput: videoDevice,
+            audioInput,
+            videoInput,
         },
         userInfo: {
             displayName,
@@ -364,8 +432,8 @@ function joinConference(audioDevice?: string, videoDevice?: string) {
         configOverwrite: {
             subject: roomName,
             startAudioOnly,
-            startWithAudioMuted: audioDevice == null,
-            startWithVideoMuted: videoDevice == null,
+            startWithAudioMuted: audioInput === null ? true : startWithAudioMuted,
+            startWithVideoMuted: videoInput === null ? true : startWithVideoMuted,
             // Request some log levels for inclusion in rageshakes
             // Ideally we would capture all possible log levels, but this can
             // cause Jitsi Meet to try to post various circular data structures
@@ -374,85 +442,112 @@ function joinConference(audioDevice?: string, videoDevice?: string) {
             apiLogLevels: ["warn", "error"],
         } as any,
         jwt: jwt,
+        lang: mapLanguage(normalizeLanguage(language)),
     };
 
     // Video channel widgets need some more tailored config options
     if (isVideoChannel) {
-        // Ensure that we skip Jitsi Meet's native prejoin screen, for
-        // deployments that have it enabled
-        options.configOverwrite.prejoinConfig = { enabled: false };
+        // We don't skip jitsi's prejoin screen for video rooms.
+        options.configOverwrite!.prejoinConfig = { enabled: true };
         // Use a simplified set of toolbar buttons
-        options.configOverwrite.toolbarButtons = [
-            "microphone", "camera", "desktop", "tileview", "hangup",
-        ];
+        options.configOverwrite!.toolbarButtons = ["microphone", "camera", "tileview", "hangup"];
+        // Note: We can hide the screenshare button in video rooms but not in
+        // normal conference calls, since in video rooms we control exactly what
+        // set of controls appear, but in normal calls we need to leave that up
+        // to the deployment's configuration.
+        // https://github.com/element-hq/element-web/issues/4880#issuecomment-940002464
+        if (supportsScreensharing) options.configOverwrite!.toolbarButtons.splice(2, 0, "desktop");
         // Hide all top bar elements
-        options.configOverwrite.conferenceInfo = { autoHide: [] };
+        options.configOverwrite!.conferenceInfo = { autoHide: [] };
+        // Remove the ability to hide your own tile, since we're hiding the
+        // settings button which would be the only way to get it back
+        options.configOverwrite!.disableSelfViewSettings = true;
     }
 
     meetApi = new JitsiMeetExternalAPI(jitsiDomain, options);
 
     // fires once when user joins the conference
     // (regardless of video on or off)
-    meetApi.on("videoConferenceJoined", () => {
-        if (avatarUrl) meetApi.executeCommand("avatarUrl", avatarUrl);
+    meetApi.on("videoConferenceJoined", onVideoConferenceJoined);
+    meetApi.on("videoConferenceLeft", onVideoConferenceLeft);
+    meetApi.on("readyToClose", closeConference as ExternalAPIEventCallbacks["readyToClose"]);
+    meetApi.on("errorOccurred", onErrorOccurred);
+    meetApi.on("audioMuteStatusChanged", onMuteStatusChanged);
+    meetApi.on("videoMuteStatusChanged", onVideoMuteStatusChanged);
 
-        if (widgetApi) {
-            // ignored promise because we don't care if it works
-            // noinspection JSIgnoredPromiseFromCall
-            widgetApi.setAlwaysOnScreen(true);
-            widgetApi.transport.send(ElementWidgetActions.JoinCall, {});
-        }
-
-        // Video rooms should start in tile mode
-        if (isVideoChannel) meetApi.executeCommand("setTileView", true);
-    });
-
-    meetApi.on("videoConferenceLeft", () => {
-        notifyHangup();
-        meetApi = null;
-    });
-
-    meetApi.on("readyToClose", closeConference);
-
-    meetApi.on("errorOccurred", ({ error }) => {
-        if (error.isFatal) {
-            // We got disconnected. Since Jitsi Meet might send us back to the
-            // prejoin screen, we're forced to act as if we hung up entirely.
-            notifyHangup(error.message);
-            meetApi = null;
-            closeConference();
-        }
-    });
-
-    meetApi.on("audioMuteStatusChanged", ({ muted }) => {
-        const action = muted ? ElementWidgetActions.MuteAudio : ElementWidgetActions.UnmuteAudio;
-        widgetApi?.transport.send(action, {});
-    });
-
-    meetApi.on("videoMuteStatusChanged", ({ muted }) => {
-        if (muted) {
-            // Jitsi Meet always sends a "video muted" event directly before
-            // hanging up, which we need to ignore by padding the timeout here,
-            // otherwise the React SDK will mistakenly think the user turned off
-            // their video by hand
-            setTimeout(() => {
-                if (meetApi) widgetApi?.transport.send(ElementWidgetActions.MuteVideo, {});
-            }, 200);
-        } else {
-            widgetApi?.transport.send(ElementWidgetActions.UnmuteVideo, {});
-        }
-    });
-
-    ["videoConferenceJoined", "participantJoined", "participantLeft"].forEach(event => {
-        meetApi.on(event, () => {
-            widgetApi?.transport.send(ElementWidgetActions.CallParticipants, {
-                participants: meetApi.getParticipantsInfo(),
-            });
-        });
+    (["videoConferenceJoined", "participantJoined", "participantLeft"] as const).forEach((event) => {
+        meetApi!.on(event, updateParticipants);
     });
 
     // Patch logs into rageshakes
-    meetApi.on("log", ({ logLevel, args }) =>
-        (parent as unknown as typeof global).mx_rage_logger?.log(logLevel, ...args),
-    );
+    meetApi.on("log", onLog);
 }
+
+const onVideoConferenceJoined = (): void => {
+    // Although we set our displayName with the userInfo option above, that
+    // option has a bug where it causes the name to be the HTML encoding of
+    // what was actually intended. So, we use the displayName command to at
+    // least ensure that the name is correct after entering the meeting.
+    // https://github.com/jitsi/jitsi-meet/issues/11664
+    // We can't just use these commands immediately after creating the
+    // iframe, because there's *another* bug where they can crash Jitsi by
+    // racing with its startup process.
+    if (displayName) meetApi?.executeCommand("displayName", displayName);
+    // This doesn't have a userInfo equivalent, so has to be set via commands
+    if (avatarUrl) meetApi?.executeCommand("avatarUrl", avatarUrl);
+
+    if (widgetApi) {
+        // ignored promise because we don't care if it works
+        // noinspection JSIgnoredPromiseFromCall
+        void widgetApi.setAlwaysOnScreen(true);
+        void widgetApi.transport.send(ElementWidgetActions.JoinCall, {});
+    }
+
+    // Video rooms should start in tile mode
+    if (isVideoChannel) meetApi?.executeCommand("setTileView", true);
+};
+
+const onVideoConferenceLeft = (): void => {
+    void notifyHangup();
+    meetApi = undefined;
+};
+
+const onErrorOccurred = ({ error }: Parameters<ExternalAPIEventCallbacks["errorOccurred"]>[0]): void => {
+    if (error.isFatal) {
+        // We got disconnected. Since Jitsi Meet might send us back to the
+        // prejoin screen, we're forced to act as if we hung up entirely.
+        void notifyHangup(error.message);
+        meetApi = undefined;
+        closeConference();
+    }
+};
+
+const onMuteStatusChanged = async (): Promise<void> => {
+    if (!meetApi) return;
+    const [audioEnabled, videoEnabled] = await checkAudioVideoEnabled();
+    void widgetApi?.transport.send(ElementWidgetActions.DeviceMute, {
+        audio_enabled: audioEnabled,
+        video_enabled: videoEnabled,
+    });
+};
+
+const onVideoMuteStatusChanged = ({ muted }: VideoMuteStatusChangedEvent): void => {
+    if (muted) {
+        // Jitsi Meet always sends a "video muted" event directly before
+        // hanging up, which we need to ignore by padding the timeout here,
+        // otherwise the React SDK will mistakenly think the user turned off
+        // their video by hand
+        setTimeout(() => onMuteStatusChanged, 200);
+    } else {
+        void onMuteStatusChanged();
+    }
+};
+
+const updateParticipants = (): void => {
+    void widgetApi?.transport.send(ElementWidgetActions.CallParticipants, {
+        participants: meetApi?.getParticipantsInfo(),
+    });
+};
+
+const onLog = ({ logLevel, args }: LogEvent): void =>
+    (parent as unknown as typeof global).mx_rage_logger?.log(logLevel, ...args);
