@@ -1,84 +1,87 @@
 /*
+Copyright 2024 New Vector Ltd.
+Copyright 2022 Šimon Brandner <simon.bra.ag@gmail.com>
+Copyright 2018-2021 New Vector Ltd
+Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
 Copyright 2016 Aviral Dasgupta
 Copyright 2016 OpenMarket Ltd
-Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
-Copyright 2018 - 2021 New Vector Ltd
-Copyright 2022 Šimon Brandner <simon.bra.ag@gmail.com>
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
 */
 
-import { UpdateCheckStatus } from "matrix-react-sdk/src/BasePlatform";
-import BaseEventIndexManager, {
-    ICrawlerCheckpoint,
-    IEventAndProfile,
-    IIndexStats,
-    ISearchArgs,
-} from 'matrix-react-sdk/src/indexing/BaseEventIndexManager';
-import dis from 'matrix-react-sdk/src/dispatcher/dispatcher';
-import { _t } from 'matrix-react-sdk/src/languageHandler';
-import SdkConfig from 'matrix-react-sdk/src/SdkConfig';
-import { IConfigOptions } from "matrix-react-sdk/src/IConfigOptions";
-import * as rageshake from 'matrix-react-sdk/src/rageshake/rageshake';
-import { MatrixClient } from "matrix-js-sdk/src/client";
-import { Room } from "matrix-js-sdk/src/models/room";
-import Modal from "matrix-react-sdk/src/Modal";
-import InfoDialog from "matrix-react-sdk/src/components/views/dialogs/InfoDialog";
-import Spinner from "matrix-react-sdk/src/components/views/elements/Spinner";
+import {
+    type MatrixClient,
+    type Room,
+    type MatrixEvent,
+    type OidcRegistrationClientMetadata,
+} from "matrix-js-sdk/src/matrix";
 import React from "react";
-import { randomString } from "matrix-js-sdk/src/randomstring";
-import { Action } from "matrix-react-sdk/src/dispatcher/actions";
-import { ActionPayload } from "matrix-react-sdk/src/dispatcher/payloads";
-import { showToast as showUpdateToast } from "matrix-react-sdk/src/toasts/UpdateToast";
-import { CheckUpdatesPayload } from "matrix-react-sdk/src/dispatcher/payloads/CheckUpdatesPayload";
-import ToastStore from "matrix-react-sdk/src/stores/ToastStore";
-import GenericExpiringToast from "matrix-react-sdk/src/components/views/toasts/GenericExpiringToast";
-import { IMatrixProfile, IEventWithRoomId as IMatrixEvent, IResultRoomEvents } from "matrix-js-sdk/src/@types/search";
+import { secureRandomString } from "matrix-js-sdk/src/randomstring";
 import { logger } from "matrix-js-sdk/src/logger";
-import { MatrixEvent } from "matrix-js-sdk/src/models/event";
 
-import VectorBasePlatform from './VectorBasePlatform';
+import BasePlatform, { UpdateCheckStatus, type UpdateStatus } from "../../BasePlatform";
+import type BaseEventIndexManager from "../../indexing/BaseEventIndexManager";
+import dis from "../../dispatcher/dispatcher";
+import SdkConfig from "../../SdkConfig";
+import { type IConfigOptions } from "../../IConfigOptions";
+import * as rageshake from "../../rageshake/rageshake";
+import Modal from "../../Modal";
+import InfoDialog from "../../components/views/dialogs/InfoDialog";
+import Spinner from "../../components/views/elements/Spinner";
+import { Action } from "../../dispatcher/actions";
+import { type ActionPayload } from "../../dispatcher/payloads";
+import { showToast as showUpdateToast } from "../../toasts/UpdateToast";
+import { type CheckUpdatesPayload } from "../../dispatcher/payloads/CheckUpdatesPayload";
+import ToastStore from "../../stores/ToastStore";
+import GenericExpiringToast from "../../components/views/toasts/GenericExpiringToast";
+import { BreadcrumbsStore } from "../../stores/BreadcrumbsStore";
+import { UPDATE_EVENT } from "../../stores/AsyncStore";
+import { avatarUrlForRoom, getInitialLetter } from "../../Avatar";
+import DesktopCapturerSourcePicker from "../../components/views/elements/DesktopCapturerSourcePicker";
+import { MatrixClientPeg } from "../../MatrixClientPeg";
+import { SeshatIndexManager } from "./SeshatIndexManager";
+import { IPCManager } from "./IPCManager";
+import { _t } from "../../languageHandler";
 
-const electron = window.electron;
-const isMac = navigator.platform.toUpperCase().includes('MAC');
+interface SquirrelUpdate {
+    releaseNotes: string;
+    releaseName: string;
+    releaseDate: Date;
+    updateURL: string;
+}
+
+const SSO_ID_KEY = "element-desktop-ssoid";
+
+const isMac = navigator.platform.toUpperCase().includes("MAC");
 
 function platformFriendlyName(): string {
     // used to use window.process but the same info is available here
-    if (navigator.userAgent.includes('Macintosh')) {
-        return 'macOS';
-    } else if (navigator.userAgent.includes('FreeBSD')) {
-        return 'FreeBSD';
-    } else if (navigator.userAgent.includes('OpenBSD')) {
-        return 'OpenBSD';
-    } else if (navigator.userAgent.includes('SunOS')) {
-        return 'SunOS';
-    } else if (navigator.userAgent.includes('Windows')) {
-        return 'Windows';
-    } else if (navigator.userAgent.includes('Linux')) {
-        return 'Linux';
+    if (navigator.userAgent.includes("Macintosh")) {
+        return "macOS";
+    } else if (navigator.userAgent.includes("FreeBSD")) {
+        return "FreeBSD";
+    } else if (navigator.userAgent.includes("OpenBSD")) {
+        return "OpenBSD";
+    } else if (navigator.userAgent.includes("SunOS")) {
+        return "SunOS";
+    } else if (navigator.userAgent.includes("Windows")) {
+        return "Windows";
+    } else if (navigator.userAgent.includes("Linux")) {
+        return "Linux";
     } else {
-        return 'Unknown';
+        return "Unknown";
     }
 }
 
-function _onAction(payload: ActionPayload) {
+function onAction(payload: ActionPayload): void {
     // Whitelist payload actions, no point sending most across
-    if (['call_state'].includes(payload.action)) {
-        electron.send('app_onAction', payload);
+    if (["call_state"].includes(payload.action)) {
+        window.electron!.send("app_onAction", payload);
     }
 }
 
-function getUpdateCheckStatus(status: boolean | string) {
+function getUpdateCheckStatus(status: boolean | string): UpdateStatus {
     if (status === true) {
         return { status: UpdateCheckStatus.Downloading };
     } else if (status === false) {
@@ -91,185 +94,83 @@ function getUpdateCheckStatus(status: boolean | string) {
     }
 }
 
-interface IPCPayload {
-    id?: number;
-    error?: string;
-    reply?: any;
-}
-
-class SeshatIndexManager extends BaseEventIndexManager {
-    private pendingIpcCalls: Record<number, { resolve, reject }> = {};
-    private nextIpcCallId = 0;
-
-    constructor() {
-        super();
-
-        electron.on('seshatReply', this.onIpcReply);
-    }
-
-    private async ipcCall(name: string, ...args: any[]): Promise<any> {
-        // TODO this should be moved into the preload.js file.
-        const ipcCallId = ++this.nextIpcCallId;
-        return new Promise((resolve, reject) => {
-            this.pendingIpcCalls[ipcCallId] = { resolve, reject };
-            window.electron.send('seshat', { id: ipcCallId, name, args });
-        });
-    }
-
-    private onIpcReply = (ev: {}, payload: IPCPayload) => {
-        if (payload.id === undefined) {
-            logger.warn("Ignoring IPC reply with no ID");
-            return;
-        }
-
-        if (this.pendingIpcCalls[payload.id] === undefined) {
-            logger.warn("Unknown IPC payload ID: " + payload.id);
-            return;
-        }
-
-        const callbacks = this.pendingIpcCalls[payload.id];
-        delete this.pendingIpcCalls[payload.id];
-        if (payload.error) {
-            callbacks.reject(payload.error);
-        } else {
-            callbacks.resolve(payload.reply);
-        }
-    };
-
-    async supportsEventIndexing(): Promise<boolean> {
-        return this.ipcCall('supportsEventIndexing');
-    }
-
-    async initEventIndex(userId: string, deviceId: string): Promise<void> {
-        return this.ipcCall('initEventIndex', userId, deviceId);
-    }
-
-    async addEventToIndex(ev: IMatrixEvent, profile: IMatrixProfile): Promise<void> {
-        return this.ipcCall('addEventToIndex', ev, profile);
-    }
-
-    async deleteEvent(eventId: string): Promise<boolean> {
-        return this.ipcCall('deleteEvent', eventId);
-    }
-
-    async isEventIndexEmpty(): Promise<boolean> {
-        return this.ipcCall('isEventIndexEmpty');
-    }
-
-    async isRoomIndexed(roomId: string): Promise<boolean> {
-        return this.ipcCall('isRoomIndexed', roomId);
-    }
-
-    async commitLiveEvents(): Promise<void> {
-        return this.ipcCall('commitLiveEvents');
-    }
-
-    async searchEventIndex(searchConfig: ISearchArgs): Promise<IResultRoomEvents> {
-        return this.ipcCall('searchEventIndex', searchConfig);
-    }
-
-    async addHistoricEvents(
-        events: IEventAndProfile[],
-        checkpoint: ICrawlerCheckpoint | null,
-        oldCheckpoint: ICrawlerCheckpoint | null,
-    ): Promise<boolean> {
-        return this.ipcCall('addHistoricEvents', events, checkpoint, oldCheckpoint);
-    }
-
-    async addCrawlerCheckpoint(checkpoint: ICrawlerCheckpoint): Promise<void> {
-        return this.ipcCall('addCrawlerCheckpoint', checkpoint);
-    }
-
-    async removeCrawlerCheckpoint(checkpoint: ICrawlerCheckpoint): Promise<void> {
-        return this.ipcCall('removeCrawlerCheckpoint', checkpoint);
-    }
-
-    async loadFileEvents(args): Promise<IEventAndProfile[]> {
-        return this.ipcCall('loadFileEvents', args);
-    }
-
-    async loadCheckpoints(): Promise<ICrawlerCheckpoint[]> {
-        return this.ipcCall('loadCheckpoints');
-    }
-
-    async closeEventIndex(): Promise<void> {
-        return this.ipcCall('closeEventIndex');
-    }
-
-    async getStats(): Promise<IIndexStats> {
-        return this.ipcCall('getStats');
-    }
-
-    async getUserVersion(): Promise<number> {
-        return this.ipcCall('getUserVersion');
-    }
-
-    async setUserVersion(version: number): Promise<void> {
-        return this.ipcCall('setUserVersion', version);
-    }
-
-    async deleteEventIndex(): Promise<void> {
-        return this.ipcCall('deleteEventIndex');
-    }
-}
-
-export default class ElectronPlatform extends VectorBasePlatform {
-    private eventIndexManager: BaseEventIndexManager = new SeshatIndexManager();
-    private pendingIpcCalls: Record<number, { resolve, reject }> = {};
-    private nextIpcCallId = 0;
+export default class ElectronPlatform extends BasePlatform {
+    private readonly ipc = new IPCManager("ipcCall", "ipcReply");
+    private readonly eventIndexManager: BaseEventIndexManager = new SeshatIndexManager();
     // this is the opaque token we pass to the HS which when we get it in our callback we can resolve to a profile
-    private ssoID: string = randomString(32);
+    private readonly ssoID: string = secureRandomString(32);
 
-    constructor() {
+    public constructor() {
         super();
 
-        dis.register(_onAction);
+        if (!window.electron) {
+            throw new Error("Cannot instantiate ElectronPlatform, window.electron is not set");
+        }
+
+        dis.register(onAction);
         /*
             IPC Call `check_updates` returns:
             true if there is an update available
             false if there is not
             or the error if one is encountered
          */
-        electron.on('check_updates', (event, status) => {
+        window.electron.on("check_updates", (event, status) => {
             dis.dispatch<CheckUpdatesPayload>({
                 action: Action.CheckUpdates,
                 ...getUpdateCheckStatus(status),
             });
         });
 
+        // `userAccessToken` (IPC) is requested by the main process when appending authentication
+        // to media downloads. A reply is sent over the same channel.
+        window.electron.on("userAccessToken", () => {
+            window.electron!.send("userAccessToken", MatrixClientPeg.get()?.getAccessToken());
+        });
+
+        // `homeserverUrl` (IPC) is requested by the main process. A reply is sent over the same channel.
+        window.electron.on("homeserverUrl", () => {
+            window.electron!.send("homeserverUrl", MatrixClientPeg.get()?.getHomeserverUrl());
+        });
+
+        // `serverSupportedVersions` is requested by the main process when it needs to know if the
+        // server supports a particular version. This is primarily used to detect authenticated media
+        // support. A reply is sent over the same channel.
+        window.electron.on("serverSupportedVersions", async () => {
+            window.electron!.send("serverSupportedVersions", await MatrixClientPeg.get()?.getVersions());
+        });
+
         // try to flush the rageshake logs to indexeddb before quit.
-        electron.on('before-quit', function() {
-            logger.log('element-desktop closing');
+        window.electron.on("before-quit", function () {
+            logger.log("element-desktop closing");
             rageshake.flush();
         });
 
-        electron.on('ipcReply', this.onIpcReply);
-        electron.on('update-downloaded', this.onUpdateDownloaded);
+        window.electron.on("update-downloaded", this.onUpdateDownloaded);
 
-        electron.on('preferences', () => {
+        window.electron.on("preferences", () => {
             dis.fire(Action.ViewUserSettings);
         });
 
-        electron.on('userDownloadCompleted', (ev, { id, name }) => {
+        window.electron.on("userDownloadCompleted", (ev, { id, name }) => {
             const key = `DOWNLOAD_TOAST_${id}`;
 
-            const onAccept = () => {
-                electron.send('userDownloadAction', { id, open: true });
+            const onAccept = (): void => {
+                window.electron!.send("userDownloadAction", { id, open: true });
                 ToastStore.sharedInstance().dismissToast(key);
             };
 
-            const onDismiss = () => {
-                electron.send('userDownloadAction', { id });
+            const onDismiss = (): void => {
+                window.electron!.send("userDownloadAction", { id });
             };
 
             ToastStore.sharedInstance().addOrReplaceToast({
                 key,
-                title: _t("Download Completed"),
+                title: _t("download_completed"),
                 props: {
                     description: name,
-                    acceptLabel: _t("Open"),
-                    onAccept,
-                    dismissLabel: _t("Dismiss"),
+                    primaryLabel: _t("action|open"),
+                    onPrimaryClick: onAccept,
+                    dismissLabel: _t("action|dismiss"),
                     onDismiss,
                     numSeconds: 10,
                 },
@@ -278,14 +179,37 @@ export default class ElectronPlatform extends VectorBasePlatform {
             });
         });
 
-        this.ipcCall("startSSOFlow", this.ssoID);
+        window.electron.on("openDesktopCapturerSourcePicker", async () => {
+            const { finished } = Modal.createDialog(DesktopCapturerSourcePicker);
+            const [source] = await finished;
+            // getDisplayMedia promise does not return if no dummy is passed here as source
+            await this.ipc.call("callDisplayMediaCallback", source ?? { id: "", name: "", thumbnailURL: "" });
+        });
+
+        void this.ipc.call("startSSOFlow", this.ssoID);
+
+        BreadcrumbsStore.instance.on(UPDATE_EVENT, this.onBreadcrumbsUpdate);
     }
 
-    async getConfig(): Promise<IConfigOptions> {
-        return this.ipcCall('getConfig');
+    public async getConfig(): Promise<IConfigOptions | undefined> {
+        return this.ipc.call("getConfig");
     }
 
-    onUpdateDownloaded = async (ev, { releaseNotes, releaseName }) => {
+    private onBreadcrumbsUpdate = (): void => {
+        const rooms = BreadcrumbsStore.instance.rooms.slice(0, 7).map((r) => ({
+            roomId: r.roomId,
+            avatarUrl: avatarUrlForRoom(
+                r,
+                Math.floor(60 * window.devicePixelRatio),
+                Math.floor(60 * window.devicePixelRatio),
+                "crop",
+            ),
+            initial: getInitialLetter(r.name),
+        }));
+        void this.ipc.call("breadcrumbs", rooms);
+    };
+
+    private onUpdateDownloaded = async (ev: Event, { releaseNotes, releaseName }: SquirrelUpdate): Promise<void> => {
         dis.dispatch<CheckUpdatesPayload>({
             action: Action.CheckUpdates,
             status: UpdateCheckStatus.Ready,
@@ -295,17 +219,15 @@ export default class ElectronPlatform extends VectorBasePlatform {
         }
     };
 
-    getHumanReadableName(): string {
-        return 'Electron Platform'; // no translation required: only used for analytics
+    public getHumanReadableName(): string {
+        return "Electron Platform"; // no translation required: only used for analytics
     }
 
     /**
      * Return true if platform supports multi-language
      * spell-checking, otherwise false.
      */
-    supportsMultiLanguageSpellCheck(): boolean {
-        // Electron uses OS spell checking on macOS, so no need for in-app options
-        if (isMac) return false;
+    public supportsSpellCheckSettings(): boolean {
         return true;
     }
 
@@ -317,259 +239,249 @@ export default class ElectronPlatform extends VectorBasePlatform {
         if (this.notificationCount === count) return;
         super.setNotificationCount(count);
 
-        electron.send('setBadgeCount', count);
+        window.electron!.send("setBadgeCount", count);
     }
 
-    supportsNotifications(): boolean {
+    public supportsNotifications(): boolean {
         return true;
     }
 
-    maySendNotifications(): boolean {
+    public maySendNotifications(): boolean {
         return true;
     }
 
-    displayNotification(title: string, msg: string, avatarUrl: string, room: Room, ev?: MatrixEvent): Notification {
+    public displayNotification(
+        title: string,
+        msg: string,
+        avatarUrl: string,
+        room: Room,
+        ev?: MatrixEvent,
+    ): Notification {
         // GNOME notification spec parses HTML tags for styling...
         // Electron Docs state all supported linux notification systems follow this markup spec
         // https://github.com/electron/electron/blob/master/docs/tutorial/desktop-environment-integration.md#linux
         // maybe we should pass basic styling (italics, bold, underline) through from MD
         // we only have to strip out < and > as the spec doesn't include anything about things like &amp;
         // so we shouldn't assume that all implementations will treat those properly. Very basic tag parsing is done.
-        if (navigator.userAgent.includes('Linux')) {
-            msg = msg.replace(/</g, '&lt;').replace(/>/g, '&gt;');
+        if (navigator.userAgent.includes("Linux")) {
+            msg = msg.replace(/</g, "&lt;").replace(/>/g, "&gt;");
         }
 
-        const notification = super.displayNotification(
-            title,
-            msg,
-            avatarUrl,
-            room,
-            ev,
-        );
+        const notification = super.displayNotification(title, msg, avatarUrl, room, ev);
 
-        const handler = notification.onclick as Function;
-        notification.onclick = () => {
+        const handler = notification.onclick as () => void;
+        notification.onclick = (): void => {
             handler?.();
-            this.ipcCall('focusWindow');
+            void this.ipc.call("focusWindow");
         };
 
         return notification;
     }
 
-    loudNotification(ev: MatrixEvent, room: Room) {
-        electron.send('loudNotification');
+    public loudNotification(ev: MatrixEvent, room: Room): void {
+        window.electron!.send("loudNotification");
     }
 
-    async getAppVersion(): Promise<string> {
-        return this.ipcCall('getAppVersion');
-    }
-
-    supportsAutoLaunch(): boolean {
+    public needsUrlTooltips(): boolean {
         return true;
     }
 
-    async getAutoLaunchEnabled(): Promise<boolean> {
-        return this.ipcCall('getAutoLaunchEnabled');
+    public async getAppVersion(): Promise<string> {
+        return this.ipc.call("getAppVersion");
     }
 
-    async setAutoLaunchEnabled(enabled: boolean): Promise<void> {
-        return this.ipcCall('setAutoLaunchEnabled', enabled);
+    public supportsSetting(settingName?: string): boolean {
+        switch (settingName) {
+            case "Electron.showTrayIcon": // Things other than Mac support tray icons
+            case "Electron.alwaysShowMenuBar": // This isn't relevant on Mac as Menu bars don't live in the app window
+                return !isMac;
+            default:
+                return true;
+        }
     }
 
-    supportsWarnBeforeExit(): boolean {
-        return true;
+    public getSettingValue(settingName: string): Promise<any> {
+        return this.ipc.call("getSettingValue", settingName);
     }
 
-    async shouldWarnBeforeExit(): Promise<boolean> {
-        return this.ipcCall('shouldWarnBeforeExit');
+    public setSettingValue(settingName: string, value: any): Promise<void> {
+        return this.ipc.call("setSettingValue", settingName, value);
     }
 
-    async setWarnBeforeExit(enabled: boolean): Promise<void> {
-        return this.ipcCall('setWarnBeforeExit', enabled);
-    }
-
-    supportsAutoHideMenuBar(): boolean {
-        // This is irelevant on Mac as Menu bars don't live in the app window
-        return !isMac;
-    }
-
-    async getAutoHideMenuBarEnabled(): Promise<boolean> {
-        return this.ipcCall('getAutoHideMenuBarEnabled');
-    }
-
-    async setAutoHideMenuBarEnabled(enabled: boolean): Promise<void> {
-        return this.ipcCall('setAutoHideMenuBarEnabled', enabled);
-    }
-
-    supportsMinimizeToTray(): boolean {
-        // Things other than Mac support tray icons
-        return !isMac;
-    }
-
-    async getMinimizeToTrayEnabled(): Promise<boolean> {
-        return this.ipcCall('getMinimizeToTrayEnabled');
-    }
-
-    async setMinimizeToTrayEnabled(enabled: boolean): Promise<void> {
-        return this.ipcCall('setMinimizeToTrayEnabled', enabled);
-    }
-
-    public supportsTogglingHardwareAcceleration(): boolean {
-        return true;
-    }
-
-    public async getHardwareAccelerationEnabled(): Promise<boolean> {
-        return this.ipcCall('getHardwareAccelerationEnabled');
-    }
-
-    public async setHardwareAccelerationEnabled(enabled: boolean): Promise<void> {
-        return this.ipcCall('setHardwareAccelerationEnabled', enabled);
-    }
-
-    async canSelfUpdate(): Promise<boolean> {
-        const feedUrl = await this.ipcCall('getUpdateFeedUrl');
+    public async canSelfUpdate(): Promise<boolean> {
+        const feedUrl = await this.ipc.call("getUpdateFeedUrl");
         return Boolean(feedUrl);
     }
 
-    startUpdateCheck() {
+    public startUpdateCheck(): void {
         super.startUpdateCheck();
-        electron.send('check_updates');
+        window.electron!.send("check_updates");
     }
 
-    installUpdate() {
+    public installUpdate(): void {
         // IPC to the main process to install the update, since quitAndInstall
         // doesn't fire the before-quit event so the main process needs to know
         // it should exit.
-        electron.send('install_update');
+        window.electron!.send("install_update");
     }
 
-    getDefaultDeviceDisplayName(): string {
+    public getDefaultDeviceDisplayName(): string {
         const brand = SdkConfig.get().brand;
-        return _t('%(brand)s Desktop (%(platformName)s)', {
+        return _t("desktop_default_device_name", {
             brand,
             platformName: platformFriendlyName(),
         });
     }
 
-    screenCaptureErrorString(): string | null {
-        return null;
+    public requestNotificationPermission(): Promise<string> {
+        return Promise.resolve("granted");
     }
 
-    requestNotificationPermission(): Promise<string> {
-        return Promise.resolve('granted');
-    }
-
-    reload() {
+    public reload(): void {
         window.location.reload();
     }
 
-    private async ipcCall(name: string, ...args: any[]): Promise<any> {
-        const ipcCallId = ++this.nextIpcCallId;
-        return new Promise((resolve, reject) => {
-            this.pendingIpcCalls[ipcCallId] = { resolve, reject };
-            window.electron.send('ipcCall', { id: ipcCallId, name, args });
-            // Maybe add a timeout to these? Probably not necessary.
-        });
-    }
-
-    private onIpcReply = (ev, payload) => {
-        if (payload.id === undefined) {
-            logger.warn("Ignoring IPC reply with no ID");
-            return;
-        }
-
-        if (this.pendingIpcCalls[payload.id] === undefined) {
-            logger.warn("Unknown IPC payload ID: " + payload.id);
-            return;
-        }
-
-        const callbacks = this.pendingIpcCalls[payload.id];
-        delete this.pendingIpcCalls[payload.id];
-        if (payload.error) {
-            callbacks.reject(payload.error);
-        } else {
-            callbacks.resolve(payload.reply);
-        }
-    };
-
-    getEventIndexingManager(): BaseEventIndexManager | null {
+    public getEventIndexingManager(): BaseEventIndexManager | null {
         return this.eventIndexManager;
     }
 
-    async setLanguage(preferredLangs: string[]) {
-        return this.ipcCall('setLanguage', preferredLangs);
+    public async setLanguage(preferredLangs: string[]): Promise<any> {
+        return this.ipc.call("setLanguage", preferredLangs);
     }
 
-    setSpellCheckLanguages(preferredLangs: string[]) {
-        this.ipcCall('setSpellCheckLanguages', preferredLangs).catch(error => {
+    public setSpellCheckEnabled(enabled: boolean): void {
+        this.ipc.call("setSpellCheckEnabled", enabled).catch((error) => {
+            logger.log("Failed to send setSpellCheckEnabled IPC to Electron");
+            logger.error(error);
+        });
+    }
+
+    public async getSpellCheckEnabled(): Promise<boolean> {
+        return this.ipc.call("getSpellCheckEnabled");
+    }
+
+    public setSpellCheckLanguages(preferredLangs: string[]): void {
+        this.ipc.call("setSpellCheckLanguages", preferredLangs).catch((error) => {
             logger.log("Failed to send setSpellCheckLanguages IPC to Electron");
             logger.error(error);
         });
     }
 
-    async getSpellCheckLanguages(): Promise<string[]> {
-        return this.ipcCall('getSpellCheckLanguages');
+    public async getSpellCheckLanguages(): Promise<string[]> {
+        return this.ipc.call("getSpellCheckLanguages");
     }
 
-    async getDesktopCapturerSources(options: GetSourcesOptions): Promise<Array<DesktopCapturerSource>> {
-        return this.ipcCall('getDesktopCapturerSources', options);
+    public async getDesktopCapturerSources(options: GetSourcesOptions): Promise<Array<DesktopCapturerSource>> {
+        return this.ipc.call("getDesktopCapturerSources", options);
     }
 
-    supportsDesktopCapturer(): boolean {
+    public supportsDesktopCapturer(): boolean {
         return true;
     }
 
-    async getAvailableSpellCheckLanguages(): Promise<string[]> {
-        return this.ipcCall('getAvailableSpellCheckLanguages');
+    public supportsJitsiScreensharing(): boolean {
+        // See https://github.com/element-hq/element-web/issues/4880
+        return false;
     }
 
-    getSSOCallbackUrl(fragmentAfterLogin: string): URL {
+    public async getAvailableSpellCheckLanguages(): Promise<string[]> {
+        return this.ipc.call("getAvailableSpellCheckLanguages");
+    }
+
+    public getSSOCallbackUrl(fragmentAfterLogin?: string): URL {
         const url = super.getSSOCallbackUrl(fragmentAfterLogin);
         url.protocol = "element";
-        url.searchParams.set("element-desktop-ssoid", this.ssoID);
+        url.searchParams.set(SSO_ID_KEY, this.ssoID);
         return url;
     }
 
-    startSingleSignOn(mxClient: MatrixClient, loginType: "sso" | "cas", fragmentAfterLogin: string, idpId?: string) {
+    public startSingleSignOn(
+        mxClient: MatrixClient,
+        loginType: "sso" | "cas",
+        fragmentAfterLogin: string,
+        idpId?: string,
+    ): void {
         // this will get intercepted by electron-main will-navigate
         super.startSingleSignOn(mxClient, loginType, fragmentAfterLogin, idpId);
-        Modal.createTrackedDialog('Electron', 'SSO', InfoDialog, {
-            title: _t("Go to your browser to complete Sign In"),
+        Modal.createDialog(InfoDialog, {
+            title: _t("auth|sso_complete_in_browser_dialog_title"),
             description: <Spinner />,
         });
     }
 
     public navigateForwardBack(back: boolean): void {
-        this.ipcCall(back ? "navigateBack" : "navigateForward");
+        void this.ipc.call(back ? "navigateBack" : "navigateForward");
     }
 
     public overrideBrowserShortcuts(): boolean {
         return true;
     }
 
-    async getPickleKey(userId: string, deviceId: string): Promise<string | null> {
+    public async getPickleKey(userId: string, deviceId: string): Promise<string | null> {
         try {
-            return await this.ipcCall('getPickleKey', userId, deviceId);
-        } catch (e) {
+            return await this.ipc.call("getPickleKey", userId, deviceId);
+        } catch {
             // if we can't connect to the password storage, assume there's no
             // pickle key
             return null;
         }
     }
 
-    async createPickleKey(userId: string, deviceId: string): Promise<string | null> {
+    public async createPickleKey(userId: string, deviceId: string): Promise<string | null> {
         try {
-            return await this.ipcCall('createPickleKey', userId, deviceId);
-        } catch (e) {
+            return await this.ipc.call("createPickleKey", userId, deviceId);
+        } catch {
             // if we can't connect to the password storage, assume there's no
             // pickle key
             return null;
         }
     }
 
-    async destroyPickleKey(userId: string, deviceId: string): Promise<void> {
+    public async destroyPickleKey(userId: string, deviceId: string): Promise<void> {
         try {
-            await this.ipcCall('destroyPickleKey', userId, deviceId);
-        } catch (e) {}
+            await this.ipc.call("destroyPickleKey", userId, deviceId);
+        } catch {}
+    }
+
+    public async clearStorage(): Promise<void> {
+        try {
+            await super.clearStorage();
+            await this.ipc.call("clearStorage");
+        } catch {}
+    }
+
+    public get baseUrl(): string {
+        // This configuration is element-desktop specific so the types here do not know about it
+        return (SdkConfig.get() as unknown as Record<string, string>)["web_base_url"] ?? "https://app.element.io";
+    }
+
+    public get defaultOidcClientUri(): string {
+        // Default to element.io as our scheme `io.element.desktop` is within its scope on default MAS policies
+        return "https://element.io";
+    }
+
+    public async getOidcClientMetadata(): Promise<OidcRegistrationClientMetadata> {
+        const baseMetadata = await super.getOidcClientMetadata();
+        return {
+            ...baseMetadata,
+            applicationType: "native",
+        };
+    }
+
+    public getOidcClientState(): string {
+        return `:${SSO_ID_KEY}:${this.ssoID}`;
+    }
+
+    /**
+     * The URL to return to after a successful OIDC authentication
+     */
+    public getOidcCallbackUrl(): URL {
+        const url = super.getOidcCallbackUrl();
+        url.protocol = "io.element.desktop";
+        // Trim the double slash into a single slash to comply with https://datatracker.ietf.org/doc/html/rfc8252#section-7.1
+        if (url.href.startsWith(`${url.protocol}://`)) {
+            url.href = url.href.replace("://", ":/");
+        }
+        return url;
     }
 }
