@@ -1,17 +1,8 @@
 /*
-Copyright 2020 New Vector Ltd
+Copyright 2020-2024 New Vector Ltd.
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
 */
 
 import { logger } from "matrix-js-sdk/src/logger";
@@ -19,12 +10,12 @@ import { logger } from "matrix-js-sdk/src/logger";
 import WebPlatform from "./WebPlatform";
 
 export default class PWAPlatform extends WebPlatform {
-    setNotificationCount(count: number) {
+    public setNotificationCount(count: number): void {
         if (!navigator.setAppBadge) return super.setNotificationCount(count);
         if (this.notificationCount === count) return;
         this.notificationCount = count;
 
-        navigator.setAppBadge(count).catch(e => {
+        navigator.setAppBadge(count).catch((e) => {
             logger.error("Failed to update PWA app badge", e);
         });
     }
