@@ -1,54 +1,89 @@
 /*
+Copyright 2017-2024 New Vector Ltd.
 Copyright 2016 Aviral Dasgupta
 Copyright 2016 OpenMarket Ltd
-Copyright 2017-2020 New Vector Ltd
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
+Please see LICENSE files in the repository root for full details.
 */
 
-import { UpdateCheckStatus } from "matrix-react-sdk/src/BasePlatform";
-import request from 'browser-request';
-import dis from 'matrix-react-sdk/src/dispatcher/dispatcher';
-import { _t } from 'matrix-react-sdk/src/languageHandler';
-import { hideToast as hideUpdateToast, showToast as showUpdateToast } from "matrix-react-sdk/src/toasts/UpdateToast";
-import { Action } from "matrix-react-sdk/src/dispatcher/actions";
-import { CheckUpdatesPayload } from 'matrix-react-sdk/src/dispatcher/payloads/CheckUpdatesPayload';
-import UAParser from 'ua-parser-js';
+import UAParser from "ua-parser-js";
 import { logger } from "matrix-js-sdk/src/logger";
 
-import VectorBasePlatform from './VectorBasePlatform';
+import { MatrixClientPeg } from "../../MatrixClientPeg";
+import BasePlatform, { UpdateCheckStatus, type UpdateStatus } from "../../BasePlatform";
+import dis from "../../dispatcher/dispatcher";
+import { hideToast as hideUpdateToast, showToast as showUpdateToast } from "../../toasts/UpdateToast";
+import { Action } from "../../dispatcher/actions";
+import { type CheckUpdatesPayload } from "../../dispatcher/payloads/CheckUpdatesPayload";
 import { parseQs } from "../url_utils";
+import { _t } from "../../languageHandler";
 
 const POKE_RATE_MS = 10 * 60 * 1000; // 10 min
 
-export default class WebPlatform extends VectorBasePlatform {
-    constructor() {
+function getNormalizedAppVersion(version: string): string {
+    // if version looks like semver with leading v, strip it (matches scripts/normalize-version.sh)
+    const semVerRegex = /^v\d+.\d+.\d+(-.+)?$/;
+    if (semVerRegex.test(version)) {
+        return version.substring(1);
+    }
+    return version;
+}
+
+export default class WebPlatform extends BasePlatform {
+    private static readonly VERSION = process.env.VERSION!; // baked in by Webpack
+
+    public constructor() {
         super();
-        // Register service worker if available on this platform
-        if ('serviceWorker' in navigator) {
-            navigator.serviceWorker.register('sw.js');
+
+        // Register the service worker in the background
+        this.tryRegisterServiceWorker().catch((e) => console.error("Error registering/updating service worker:", e));
+    }
+
+    private async tryRegisterServiceWorker(): Promise<void> {
+        if (!("serviceWorker" in navigator)) {
+            return; // not available on this platform - don't try to register the service worker
+        }
+
+        // sw.js is exported by webpack, sourced from `/src/serviceworker/index.ts`
+        const registration = await navigator.serviceWorker.register("sw.js");
+        if (!registration) {
+            // Registration didn't work for some reason - assume failed and ignore.
+            // This typically happens in Jest.
+            return;
+        }
+
+        navigator.serviceWorker.addEventListener("message", this.onServiceWorkerPostMessage.bind(this));
+        await registration.update();
+    }
+
+    private onServiceWorkerPostMessage(event: MessageEvent): void {
+        try {
+            if (event.data?.["type"] === "userinfo" && event.data?.["responseKey"]) {
+                const userId = localStorage.getItem("mx_user_id");
+                const deviceId = localStorage.getItem("mx_device_id");
+                const homeserver = MatrixClientPeg.get()?.getHomeserverUrl();
+                event.source!.postMessage({
+                    responseKey: event.data["responseKey"],
+                    userId,
+                    deviceId,
+                    homeserver,
+                });
+            }
+        } catch (e) {
+            console.error("Error responding to service worker: ", e);
         }
     }
 
-    getHumanReadableName(): string {
-        return 'Web Platform'; // no translation required: only used for analytics
+    public getHumanReadableName(): string {
+        return "Web Platform"; // no translation required: only used for analytics
     }
 
     /**
      * Returns true if the platform supports displaying
      * notifications, otherwise false.
      */
-    supportsNotifications(): boolean {
+    public supportsNotifications(): boolean {
         return Boolean(window.Notification);
     }
 
@@ -56,8 +91,8 @@ export default class WebPlatform extends VectorBasePlatform {
      * Returns true if the application currently has permission
      * to display notifications. Otherwise false.
      */
-    maySendNotifications(): boolean {
-        return window.Notification.permission === 'granted';
+    public maySendNotifications(): boolean {
+        return window.Notification.permission === "granted";
     }
 
     /**
@@ -67,58 +102,36 @@ export default class WebPlatform extends VectorBasePlatform {
      * that is 'granted' if the user allowed the request or
      * 'denied' otherwise.
      */
-    requestNotificationPermission(): Promise<string> {
+    public requestNotificationPermission(): Promise<string> {
         // annoyingly, the latest spec says this returns a
         // promise, but this is only supported in Chrome 46
         // and Firefox 47, so adapt the callback API.
-        return new Promise(function(resolve, reject) {
+        return new Promise(function (resolve, reject) {
             window.Notification.requestPermission((result) => {
                 resolve(result);
-            });
+            }).catch(reject);
         });
     }
 
-    private getMostRecentVersion(): Promise<string> {
-        // We add a cachebuster to the request to make sure that we know about
-        // the most recent version on the origin server. That might not
-        // actually be the version we'd get on a reload (particularly in the
-        // presence of intermediate caching proxies), but still: we're trying
-        // to tell the user that there is a new version.
-
-        return new Promise((resolve, reject) => {
-            request(
-                {
-                    method: "GET",
-                    url: "version",
-                    qs: { cachebuster: Date.now() },
-                },
-                (err, response, body) => {
-                    if (err || response.status < 200 || response.status >= 300) {
-                        if (err === null) err = { status: response.status };
-                        reject(err);
-                        return;
-                    }
-
-                    resolve(this.getNormalizedAppVersion(body.trim()));
-                },
-            );
+    private async getMostRecentVersion(): Promise<string> {
+        const res = await fetch("version", {
+            method: "GET",
+            cache: "no-cache",
         });
-    }
 
-    getNormalizedAppVersion(version: string): string {
-        // if version looks like semver with leading v, strip it (matches scripts/normalize-version.sh)
-        const semVerRegex = /^v\d+.\d+.\d+(-.+)?$/;
-        if (semVerRegex.test(version)) {
-            return version.substring(1);
+        if (res.ok) {
+            const text = await res.text();
+            return getNormalizedAppVersion(text.trim());
         }
-        return version;
+
+        return Promise.reject({ status: res.status });
     }
 
-    getAppVersion(): Promise<string> {
-        return Promise.resolve(this.getNormalizedAppVersion(process.env.VERSION));
+    public getAppVersion(): Promise<string> {
+        return Promise.resolve(getNormalizedAppVersion(WebPlatform.VERSION));
     }
 
-    startUpdater() {
+    public startUpdater(): void {
         // Poll for an update immediately, and reload the page now if we're out of date
         // already as we've just initialised an old version of the app somehow.
         //
@@ -127,8 +140,8 @@ export default class WebPlatform extends VectorBasePlatform {
         //
         // Ideally, loading an old copy would be impossible with the
         // cache-control: nocache HTTP header set, but Firefox doesn't always obey it :/
-        console.log("startUpdater, current version is " + this.getNormalizedAppVersion(process.env.VERSION));
-        this.pollForUpdate((version: string, newVersion: string) => {
+        console.log("startUpdater, current version is " + getNormalizedAppVersion(WebPlatform.VERSION));
+        void this.pollForUpdate((version: string, newVersion: string) => {
             const query = parseQs(location);
             if (query.updated) {
                 console.log("Update reloaded but still on an old version, stopping");
@@ -147,43 +160,47 @@ export default class WebPlatform extends VectorBasePlatform {
         setInterval(() => this.pollForUpdate(showUpdateToast, hideUpdateToast), POKE_RATE_MS);
     }
 
-    async canSelfUpdate(): Promise<boolean> {
+    public async canSelfUpdate(): Promise<boolean> {
         return true;
     }
 
-    pollForUpdate = (
+    // Exported for tests
+    public pollForUpdate = (
         showUpdate: (currentVersion: string, mostRecentVersion: string) => void,
         showNoUpdate?: () => void,
-    ) => {
-        return this.getMostRecentVersion().then((mostRecentVersion) => {
-            const currentVersion = this.getNormalizedAppVersion(process.env.VERSION);
+    ): Promise<UpdateStatus> => {
+        return this.getMostRecentVersion().then(
+            (mostRecentVersion) => {
+                const currentVersion = getNormalizedAppVersion(WebPlatform.VERSION);
 
-            if (currentVersion !== mostRecentVersion) {
-                if (this.shouldShowUpdate(mostRecentVersion)) {
-                    console.log("Update available to " + mostRecentVersion + ", will notify user");
-                    showUpdate(currentVersion, mostRecentVersion);
+                if (currentVersion !== mostRecentVersion) {
+                    if (this.shouldShowUpdate(mostRecentVersion)) {
+                        console.log("Update available to " + mostRecentVersion + ", will notify user");
+                        showUpdate(currentVersion, mostRecentVersion);
+                    } else {
+                        console.log("Update available to " + mostRecentVersion + " but won't be shown");
+                    }
+                    return { status: UpdateCheckStatus.Ready };
                 } else {
-                    console.log("Update available to " + mostRecentVersion + " but won't be shown");
+                    console.log("No update available, already on " + mostRecentVersion);
+                    showNoUpdate?.();
                 }
-                return { status: UpdateCheckStatus.Ready };
-            } else {
-                console.log("No update available, already on " + mostRecentVersion);
-                showNoUpdate?.();
-            }
 
-            return { status: UpdateCheckStatus.NotAvailable };
-        }, (err) => {
-            logger.error("Failed to poll for update", err);
-            return {
-                status: UpdateCheckStatus.Error,
-                detail: err.message || err.status ? err.status.toString() : 'Unknown Error',
-            };
-        });
+                return { status: UpdateCheckStatus.NotAvailable };
+            },
+            (err) => {
+                logger.error("Failed to poll for update", err);
+                return {
+                    status: UpdateCheckStatus.Error,
+                    detail: err.message || (err.status ? err.status.toString() : "Unknown Error"),
+                };
+            },
+        );
     };
 
-    startUpdateCheck() {
+    public startUpdateCheck(): void {
         super.startUpdateCheck();
-        this.pollForUpdate(showUpdateToast, hideUpdateToast).then((updateState) => {
+        void this.pollForUpdate(showUpdateToast, hideUpdateToast).then((updateState) => {
             dis.dispatch<CheckUpdatesPayload>({
                 action: Action.CheckUpdates,
                 ...updateState,
@@ -191,11 +208,11 @@ export default class WebPlatform extends VectorBasePlatform {
         });
     }
 
-    installUpdate() {
+    public installUpdate(): void {
         window.location.reload();
     }
 
-    getDefaultDeviceDisplayName(): string {
+    public getDefaultDeviceDisplayName(): string {
         // strip query-string and fragment from uri
         const url = new URL(window.location.href);
 
@@ -210,22 +227,14 @@ export default class WebPlatform extends VectorBasePlatform {
         let osName = ua.getOS().name || "unknown OS";
         // Stylise the value from the parser to match Apple's current branding.
         if (osName === "Mac OS") osName = "macOS";
-        return _t('%(appName)s (%(browserName)s, %(osName)s)', {
+        return _t("web_default_device_name", {
             appName,
             browserName,
             osName,
         });
     }
 
-    screenCaptureErrorString(): string | null {
-        // it won't work at all if you're not on HTTPS so whine whine whine
-        if (window.location.protocol !== "https:") {
-            return _t("You need to be using HTTPS to place a screen-sharing call.");
-        }
-        return null;
-    }
-
-    reload() {
+    public reload(): void {
         window.location.reload();
     }
 }
